% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dealWithGencode.R
\name{apiAdmin_ping}
\alias{apiAdmin_ping}
\title{Title Heartbeat to check server connectivity.}
\usage{
apiAdmin_ping()
}
\value{
boolean value
}
\description{
Title Heartbeat to check server connectivity.
}
\examples{
 apiStatus <- ifelse( apiAdmin_ping()==200, "GTEx API can be accessed", "Please check your network!")
 print(apiStatus)
}
