% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{apiEbi_ping}
\alias{apiEbi_ping}
\title{Heartbeat to check EBI API server connectivity.}
\usage{
apiEbi_ping()
}
\value{
A character of fetchContent method.
}
\description{
test EBI API server and return download method.
}
\examples{
# apiEbi_ping()
}
