% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dealWithGencode.R
\name{apiRef_gene}
\alias{apiRef_gene}
\title{Title fetch reference genes by API.}
\usage{
apiRef_gene(geneId = "", gencodeVersion = "v26", genomeBuild = "GRCh38/hg38")
}
\arguments{
\item{geneId}{a character or a character vector(versioned ensemble ID or unversioned ensemble ID)}

\item{gencodeVersion}{"v26" or "v19"}

\item{genomeBuild}{"GRCh38/hg38" or "GRCh37/hg19"}
}
\value{
queried gene info table
}
\description{
Title fetch reference genes by API.
}
\examples{
 apiRef_gene(c("ENSG00000116885.18","ENSG00000222623"), "v26", "GRCh38/hg38")
}
