% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{apiRef_geneAll}
\alias{apiRef_geneAll}
\title{Title}
\usage{
apiRef_geneAll(gencodeVersion = "v26", recordPerChunk = 2000)
}
\arguments{
\item{gencodeVersion}{A character string. "v26" or "v19".}

\item{recordPerChunk}{A integer value. From 1 to 2000.}
}
\value{

}
\description{
Title
}
\examples{
/donttest{
 gencodeGeneInfoV26 <- apiRef_geneAll("v26")
 gencodeGeneInfoV19 <- apiRef_geneAll("v19")
}
}
