% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dealWithGencode.R
\name{createTissueSiteDetailMappingData}
\alias{createTissueSiteDetailMappingData}
\title{create .rds file with GTExquery_sample function}
\usage{
createTissueSiteDetailMappingData(datasetId = "gtex_v8")
}
\arguments{
\item{datasetId}{"gtex_v8" or "gtex_v7"}
}
\value{
none
}
\description{
create .rds file with GTExquery_sample function
}
