% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{ebi_datasets}
\alias{ebi_datasets}
\title{data for ebi datasets summary}
\format{
A Data.table
\describe{
\item{study_id}{A character vector}
\item{quant_method}{A character vector}
\item{sample_group}{A character vector}
\item{tissue_id}{A character vector}
\item{study_label}{A character vector}
\item{tissue_label}{A character vector}
\item{condition_label}{A character vector}
\item{dataset_id}{A character vector}
\item{sample_size}{A integer vector}

}
}
\source{
\url{https://www.ebi.ac.uk/eqtl/api-docs/#accessing-the-api} or EBIquery_allTerm
}
\description{
a dataset containing all datasets collected in EBI catalogue
}
\keyword{internal}
