% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{ebi_qtl_groups}
\alias{ebi_qtl_groups}
\title{tissue resources in the ebi eQTL catelogue}
\format{
A Data.table
\describe{
\item{qtl_group}{A character vector}
}
}
\source{
\url{https://www.ebi.ac.uk/eqtl/api-docs/#accessing-the-api}
}
\description{
A list of 101 tissue resources
}
\keyword{internal}
