% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{ebi_study_tissues}
\alias{ebi_study_tissues}
\title{data for vignette}
\format{
A Data.table
}
\source{
\url{https://www.ebi.ac.uk/eqtl/api-docs/#accessing-the-api} or EBIquery_allTerm
}
\description{
a dataset containing all study-tissues mapping relationships.
}
\keyword{internal}
