% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_gwasDF}
\alias{example_Coloc_gwasDF}
\title{data for vignette}
\format{
A Data.table
}
\source{
\url{http://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST006001-GCST007000/GCST006085/harmonised/29892016-GCST006085-EFO_0001663-build37.f.tsv.gz}
}
\description{
head 6 rows of gwasDF of GLGC_CG0276_result.txt(GCST006085)
}
\keyword{internal}
