% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_hyprcolocResultAll}
\alias{example_Coloc_hyprcolocResultAll}
\title{data for vignette}
\format{
A Data.table
}
\source{
xQTLanalyze_coloc
}
\description{
head 5 rows of hyprcoloc results from xQTLanalyze_coloc
}
\keyword{internal}
