% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_outGenes}
\alias{example_Coloc_outGenes}
\title{data for vignette}
\format{
A Data.table
\describe{
\item{genes}{A character vector}
\item{genesUpper}{A character vector}
\item{geneSymbol}{A character vector}
\item{gencodeId}{A character vector}
\item{entrezGeneId}{A numeric vector}
\item{geneType}{A numeric vector}
\item{chromosome}{A numeric vector}
\item{start}{A numeric vector}
\item{end}{A numeric vector}
\item{strand}{A numeric vector}
\item{tss}{A numeric vector}
\item{gencodeVersion}{A numeric vector}
\item{genomeBuild}{A numeric vector}
\item{description}{A numeric vector}
}
}
\source{
xQTLanalyze_coloc
}
\description{
xx
}
\keyword{internal}
