% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_sentinelSNP}
\alias{example_Coloc_sentinelSNP}
\title{data for vignette}
\format{
A Data.table
}
\source{
xQTLanalyze_getSentinelSnp
}
\description{
head 6 rows of sentinelSnpDF of xQTLanalyze_getSentinelSnp
}
\keyword{internal}
