% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_traitsAll}
\alias{example_Coloc_traitsAll}
\title{data for vignette}
\format{
A Data.table
}
\source{
xQTLanalyze_getTraits
}
\description{
head 6 rows of traitsAll of xQTLanalyze_getTraits
}
\keyword{internal}
