% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{extractGeneInfo}
\alias{extractGeneInfo}
\title{extract gene infor of specified genome from gencodeGeneInfoAllGranges}
\usage{
extractGeneInfo(gencodeGeneInfoAllGranges, genomeVersion = "V26")
}
\arguments{
\item{gencodeGeneInfoAllGranges}{from internal data}

\item{genomeVersion}{"V26" (default) or "V19"}
}
\value{
a data.table
}
\description{
extract gene infor of specified genome from gencodeGeneInfoAllGranges
}
\examples{
\donttest{
  gencodeGeneInfo <- extractGeneInfo(gencodeGeneInfoAllGranges)
}
}
