% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{fetchContent}
\alias{fetchContent}
\title{Fetch data using url by three methods}
\usage{
fetchContent(url1, method = "download", downloadMethod = "auto")
}
\arguments{
\item{url1}{A url string.}

\item{method}{Can be chosen from "download", "curl" or "GET".}

\item{downloadMethod}{The same methods from utils::download.file function.}
}
\value{
A json object.
}
\description{
Fetch data using url by three methods
}
\examples{
\donttest{
 url1 <- "https://gtexportal.org/rest/v1/association/dyneqtl?gencodeId=ENSG00000065613.13&variantId=chr11_66561248_T_C_b38&tissueSiteDetailId=Liver&datasetId=gtex_v8"
 fetchContent(url1, method="curl")
 url1 <- "https://gtexportal.org/rest/v1/admin/ping"
 fetchContent(url1, method="download")
}
}
