% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{fetchContentEbi}
\alias{fetchContentEbi}
\title{Fetch records from}
\usage{
fetchContentEbi(
  url1,
  method = "fromJSON",
  downloadMethod = "auto",
  termSize = 1000,
  termStart = 0
)
}
\arguments{
\item{url1}{A url string.}

\item{method}{Can be chosen from "download", "curl" or "GET".}

\item{termSize}{Number of records per request.}

\item{termStart}{Start position per request.}

\item{downloadMethodThe}{same methods from utils::download.file function.}
}
\value{

}
\description{
Fetch records from
}
\examples{
\donttest{
 url1 <- "https://www.ebi.ac.uk/eqtl/api/tissues/CL_0000057/associations?gene_id=ENSG00000141510"
 gtexAsoo <- fetchContentEbi(url1)
}
}
