% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{gencodeGeneInfoAllGranges}
\alias{gencodeGeneInfoAllGranges}
\title{Gene annotations (chr1-chr22).}
\format{
A GRanges object
}
\source{
\url{https://gtexportal.org/home/}
}
\description{
A dataset containing the gene information.
}
\keyword{internal}
