% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_getSentinelSnp}
\alias{GTExanalyze_getSentinelSnp}
\title{Sentinel SNPs detection in GWAS data.}
\usage{
GTExanalyze_getSentinelSnp(
  gwasDF,
  pValueThreshold = 5e-08,
  centerRange = 1e+06
)
}
\arguments{
\item{gwasDF}{GWAS data.frame}
}
\value{

}
\description{
Sentinel SNPs detection in GWAS data.
}
\examples{
\donttest{
   gwasFile <- tempfile(pattern = "file")
   gwasURL <- "https://raw.githubusercontent.com/dingruofan/exampleData/master/gwas/AD/GLGC_AD_chr1_6_Sub3.txt"
   utils::download.file(gwasURL, destfile=gwasFile)
   gwasDF <- data.table::fread(gwasFile, sep="\t")
   gwasDF <- gwasDF[, .(rsid, chr, position, P, maf)]
   sentinelSnps <- GTExanalyze_getSentinelSnp(gwasDF)

   gwasDF <- fread("D:\\\\R_project\\\\GLGC_CG0052_result.txt.gz", sep="\t")
   gwasDF <- gwasDF[,.(rsid, chr, position, `p-value`, maf)]
   sentinelSnps <- GTExanalyze_getSentinelSnp(gwasDF)
}
}
