% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_getTraits}
\alias{GTExanalyze_getTraits}
\title{Detect the genes around the sentinel SNPs}
\usage{
GTExanalyze_getTraits(
  sentinelSnpDF,
  colocRange = 1e+06,
  genomeVersion = "grch38"
)
}
\arguments{
\item{sentinelSnpDF}{A data.table. Better be the results from the function "GTExanalyze_getSentinelSnp", five columns are required, including "rsid", "chr", "position", "pValue", and "maf".}

\item{colocRange}{A integer value. Window size centered on SNP. Default: 1e6 bp}

\item{genomeVersion}{"grch38" or "grch19". Default: "grch38"}
}
\value{
A data.table object
}
\description{
Detect the genes around the sentinel SNPs
}
\examples{
\donttest{
  sentinelSnpsURL <- "https://gitee.com/stronghoney/exampleData/raw/master/gwas/GLGC_CG0052/sentinelSnpDF.txt"
  sentinelSnpDF <- data.table::fread(rawToChar(curl::curl_fetch_memory(sentinelSnpsURL)$content), sep="\t")
  traitsAll <- GTExanalyze_getTraits(sentinelSnpDF, colocRange=1e6, genomeVersion="grch37" )
}
}
