% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_traitEqtlSig}
\alias{GTExanalyze_traitEqtlSig}
\title{}
\usage{
GTExanalyze_traitEqtlSig(traitGenes, sentinelSnps, tissueSiteDetail = "")
}
\arguments{
\item{tissueSiteDetail}{}
}
\value{

}
\description{

}
\examples{
\donttest{
  traitGenes <- traitsAll$gencodeId
  sentinelSnps <- traitsAll$rsid
}
}
