% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExbiolinks.R
\docType{package}
\name{GTExbiolinks}
\alias{GTExbiolinks}
\title{GTExbiolinks package}
\description{
The functions you're likely to need from \pkg{GTExbiolinks} is
Otherwise refer to the vignettes to see
how to format the documentation.
}
