% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_aQTLAllAsso}
\alias{GTExdownload_aQTLAllAsso}
\title{Fetch all eQTL associations from EBI eQTL category.}
\usage{
GTExdownload_aQTLAllAsso(
  gene = "",
  geneType = "geneSymbol",
  refSeq = "",
  rsid = "",
  pThreshold = 1,
  variantId = "",
  tissueSiteDetail = ""
)
}
\arguments{
\item{gene}{gene symbol or gencode ID.}

\item{geneType}{A character string. "geneSymbol"(default), "gencodeId" or "geneCategory".}

\item{refSeq}{A character string.}

\item{rsid}{dbSNP ID.}

\item{variantId}{A character string. like "chr17_5294580_CT_C_b38".}

\item{tissueSiteDetail}{}
}
\value{

}
\description{
Fetch all eQTL associations from EBI eQTL category.
}
\examples{
\donttest{
  GTExdownload_aQTLAllAsso( gene="TP53", tissueSiteDetail = "Adipose - Visceral (Omentum)", pThreshold=1e-2)
  GTExdownload_aQTLAllAsso( refSeq = "NM_001291581.2", tissueSiteDetail = "Adipose - Visceral (Omentum)" )
  GTExdownload_aQTLAllAsso( rsid = "rs3026133", tissueSiteDetail = "Adipose - Visceral (Omentum)", pThreshold=1e-4 )
  GTExdownload_aQTLAllAsso( variantId = "chr17_5289661_A_G_b38", tissueSiteDetail = "Adipose - Visceral (Omentum)" )
}
}
