% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_eqtlDIY}
\alias{GTExdownload_eqtlDIY}
\title{query eQTLs that may or may not be significant}
\usage{
GTExdownload_eqtlDIY(x)
}
\arguments{
\item{x}{aa
\tabular{rrrrr}{
\strong{mpg} \tab \strong{cyl} \tab \strong{disp} \tab \strong{hp} \tab \strong{drat} \cr
21.0 \tab 6 \tab 160 \tab 110 \tab 3.90\cr
21.0 \tab 6 \tab 160 \tab 110 \tab 3.90\cr
22.8 \tab 4 \tab 108 \tab  93 \tab 3.85\cr
21.4 \tab 6 \tab 258 \tab 110 \tab 3.08\cr
18.7 \tab 8 \tab 360 \tab 175 \tab 3.15
}}
}
\value{
a data.table
}
\description{
query eQTLs that may or may not be significant
}
\examples{
\donttest{

}
}
