% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_eqtlExp}
\alias{GTExdownload_eqtlExp}
\title{Download expression data for eQTL.}
\usage{
GTExdownload_eqtlExp(
  variantName = "",
  gene = "",
  variantType = "snpId",
  geneType = "geneSymbol",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\value{

}
\description{
This function fetch normalized gene expression data for a eQTL pair.
}
