% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{GTExdownload_exp}
\alias{GTExdownload_exp}
\title{Fetch normalized gene expression.}
\usage{
GTExdownload_exp(
  genes = "",
  geneType = "geneSymbol",
  tissueSiteDetail = "Liver",
  datasetId = "gtex_v8",
  toSummarizedExperiment = TRUE,
  recordPerChunk = 150
)
}
\arguments{
\item{genes}{Following gene types are supported:
\itemize{
\item \strong{Gene symbol}.

A character string or a character vector (case ignored). like: "tp53","naDK","SDF4".
\item \strong{Gencode/ensemble id} (versioned or unversioned).
}}

\item{geneType}{A character string. Types of queried genes. Options: "geneSymbol" (default) and "gencodeId";}

\item{tissueSiteDetail}{For GTEx v8, must be chosen from the following terms:
\itemize{
\item Adipose - Subcutaneous
\item Adipose - Visceral (Omentum)
\item Adrenal Gland
\item Artery - Aorta
\item Artery - Coronary
\item Artery - Tibial
\item Bladder
\item Brain - Amygdala
\item Brain - Anterior cingulate cortex (BA24)
\item Brain - Caudate (basal ganglia)
\item Brain - Cerebellar Hemisphere
\item Brain - Cerebellum
\item Brain - Cortex
\item Brain - Frontal Cortex (BA9)
\item Brain - Hippocampus
\item Brain - Hypothalamus
\item Brain - Nucleus accumbens (basal ganglia)
\item Brain - Putamen (basal ganglia)
\item Brain - Spinal cord (cervical c-1)
\item Brain - Substantia nigra
\item Breast - Mammary Tissue
\item Cells - Cultured fibroblasts
\item Cells - EBV-transformed lymphocytes
\item Cervix - Ectocervix
\item Cervix - Endocervix
\item Colon - Sigmoid
\item Colon - Transverse
\item Esophagus - Gastroesophageal Junction
\item Esophagus - Mucosa
\item Esophagus - Muscularis
\item Fallopian Tube
\item Heart - Atrial Appendage
\item Heart - Left Ventricle
\item Kidney - Cortex
\item Kidney - Medulla
\item Liver
\item Lung
\item Minor Salivary Gland
\item Muscle - Skeletal
\item Nerve - Tibial
\item Ovary
\item Pancreas
\item Pituitary
\item Prostate
\item Skin - Not Sun Exposed (Suprapubic)
\item Skin - Sun Exposed (Lower leg)
\item Small Intestine - Terminal Ileum
\item Spleen
\item Stomach
\item Testis
\item Thyroid
\item Uterus
\item Vagina
\item Whole Blood
}
For GTEx v7, must be following terms:
\itemize{
\item Adipose - Subcutaneous
\item Adipose - Visceral (Omentum)
\item Adrenal Gland
\item Artery - Aorta
\item Artery - Coronary
\item Artery - Tibial
\item Bladder
\item Brain - Amygdala
\item Brain - Anterior cingulate cortex (BA24)
\item Brain - Caudate (basal ganglia)
\item Brain - Cerebellar Hemisphere
\item Brain - Cerebellum
\item Brain - Cortex
\item Brain - Frontal Cortex (BA9)
\item Brain - Hippocampus
\item Brain - Hypothalamus
\item Brain - Nucleus accumbens (basal ganglia)
\item Brain - Putamen (basal ganglia)
\item Brain - Spinal cord (cervical c-1)
\item Brain - Substantia nigra
\item Breast - Mammary Tissue
\item Cells - EBV-transformed lymphocytes
\item Cells - Transformed fibroblasts
\item Cervix - Ectocervix
\item Cervix - Endocervix
\item Colon - Sigmoid
\item Colon - Transverse
\item Esophagus - Gastroesophageal Junction
\item Esophagus - Mucosa
\item Esophagus - Muscularis
\item Fallopian Tube
\item Heart - Atrial Appendage
\item Heart - Left Ventricle
\item Kidney - Cortex
\item Liver
\item Lung
\item Minor Salivary Gland
\item Muscle - Skeletal
\item Nerve - Tibial
\item Ovary
\item Pancreas
\item Pituitary
\item Prostate
\item Skin - Not Sun Exposed (Suprapubic)
\item Skin - Sun Exposed (Lower leg)
\item Small Intestine - Terminal Ileum
\item Spleen
\item Stomach
\item Testis
\item Thyroid
\item Uterus
\item Vagina
\item Whole Blood
}}

\item{datasetId}{A character string. Options: "gtex_v8" (default), "gtex_v7".}

\item{toSummarizedExperiment}{whether to return a data.frame or a summarizedExperiment object. Default: TRUE, return a toSummarizedExperiment object.}

\item{recordPerChunk}{A integer value (1-2000). number of records fetched per request (default: 150).}
}
\value{
return a SummarizedExperiment or data.frame object harbing gene expression profiles.
}
\description{
This function fetch queried genes' expression profiles in a tissue at the sample level.
}
\examples{
\donttest{
  expProfiles <- GTExdownload_exp(c("ENSG00000210195.2"),
                                 "gencodeId", "Liver", "gtex_v8")
  expProfiles <- GTExdownload_exp(c("tp53","naDK","SDF4"),
                                 "geneSymbol", "Artery - Coronary", "gtex_v8",
                                 toSummarizedExperiment=TRUE)
  expProfiles <- GTExdownload_exp(c("tp53","naDK","SDF4"),
                                 "geneSymbol", "Artery - Coronary", "gtex_v7")
  # get miRNA expression:
  miRNA <- GTExquery_gene (genes="miRNA", geneType="geneCategory" )
  miRNAExp <- GTExdownload_exp(miRNA$gencodeId, geneType = "gencodeId")
  geneExp <- GTExdownload_exp(c("tp53",miRNA$geneSymbol[1:100],"naDK"), geneType = "geneSymbol")
  proT <- GTExquery_gene (genes="protein coding", geneType="geneCategory", "v26" )
  proTexp <- GTExdownload_exp(proT$geneSymbol[1:100], geneType = "geneSymbol","Lung","gtex_v7")
  }
}
