% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_geneMedExp}
\alias{GTExdownload_geneMedExp}
\title{fetch gene median expression}
\usage{
GTExdownload_geneMedExp(
  genes = "",
  geneType = "geneSymbol",
  datasetId = "gtex_v8",
  tissueSiteDetail = "",
  recordPerChunk = 150
)
}
\arguments{
\item{genes}{A character vector of gene symbol/gencode ID.}

\item{geneType}{"geneSymbol" or "gencodeID".}

\item{datasetId}{"gtex_v7" or "gtex_v8"}

\item{tissueSiteDetail}{tissue name.}

\item{recordPerChunk}{1-2000. Default: 150.}
}
\value{
A data.frame
}
\description{
fetch gene median expression
}
\examples{
\donttest{
 GTExdownload_geneMedExp(genes="TP53", "geneSymbol")
}
}
