% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_ld}
\alias{GTExdownload_ld}
\title{Fetch linkage disequilibrium data}
\usage{
GTExdownload_ld(
  gene = "",
  geneType = "geneSymbol",
  datasetId = "gtex_v8",
  recordPerChunk = 100
)
}
\arguments{
\item{gene}{A gene symbol, gencode id (versioned), or a charater string of gene type.}

\item{geneType}{A character string. "geneSymbol"(default), or "gencodeId".}

\item{datasetId}{A character string. Options: "gtex_v8" (default), "gtex_v7".}

\item{recordPerChunk}{A integer value (1-500). number of records fetched per request (default: 100).}
}
\value{

}
\description{
Fetch linkage disequilibrium data for the cis-eQTLs found associated with this gene in a specified dataset.
}
\examples{
\donttest{
 GTExdownload_ld("TP53" )
 GTExdownload_ld(gene = "DDX11", datasetId="gtex_v7")
 GTExdownload_ld(gene="ENSG00000008128.22", geneType="gencodeId")

}
}
