% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{GTExquery_exp}
\alias{GTExquery_exp}
\title{Fetch normalized gene expression.}
\usage{
GTExquery_exp(
  genes = "",
  geneType = "symbol",
  tissueSiteDetail = "Liver",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{genes}{Following gene types are supported:
\itemize{
\item \strong{Gene symbol}.

A character string or a character vector (case ignored). like: "tp53","naDK","SDF4".
\item \strong{Gencode/ensemble id} (versioned or unversioned).

A character string or a character vector (case ignored). like: "ENSG00000210195.2","ENSG00000078808","Ensg00000008130.15".
\item \strong{Entrez gene ID}.

A integer string or a integer vectors. like: 51150,5590,4509.
}}

\item{geneType}{A character string. Types of queried genes. Options: "symbol" (default), "gencodeId", "entrezId";}

\item{tissueSiteDetail}{For GTEx v8, must be following terms:
\itemize{
\item All
\item Adipose - Subcutaneous
\item Adipose - Visceral (Omentum)
\item Adrenal Gland
\item Artery - Aorta
\item Artery - Coronary
\item Artery - Tibial
\item Bladder
\item Brain - Amygdala
\item Brain - Anterior cingulate cortex (BA24)
\item Brain - Caudate (basal ganglia)
\item Brain - Cerebellar Hemisphere
\item Brain - Cerebellum
\item Brain - Cortex
\item Brain - Frontal Cortex (BA9)
\item Brain - Hippocampus
\item Brain - Hypothalamus
\item Brain - Nucleus accumbens (basal ganglia)
\item Brain - Putamen (basal ganglia)
\item Brain - Spinal cord (cervical c-1)
\item Brain - Substantia nigra
\item Breast - Mammary Tissue
\item Cells - Cultured fibroblasts
\item Cells - EBV-transformed lymphocytes
\item Cervix - Ectocervix
\item Cervix - Endocervix
\item Colon - Sigmoid
\item Colon - Transverse
\item Esophagus - Gastroesophageal Junction
\item Esophagus - Mucosa
\item Esophagus - Muscularis
\item Fallopian Tube
\item Heart - Atrial Appendage
\item Heart - Left Ventricle
\item Kidney - Cortex
\item Kidney - Medulla
\item Liver
\item Lung
\item Minor Salivary Gland
\item Muscle - Skeletal
\item Nerve - Tibial
\item Ovary
\item Pancreas
\item Pituitary
\item Prostate
\item Skin - Not Sun Exposed (Suprapubic)
\item Skin - Sun Exposed (Lower leg)
\item Small Intestine - Terminal Ileum
\item Spleen
\item Stomach
\item Testis
\item Thyroid
\item Uterus
\item Vagina
\item Whole Blood
}
For GTEx v7, must be following terms:
\itemize{
\item All
\item Adipose - Subcutaneous
\item Adipose - Visceral (Omentum)
\item Adrenal Gland
\item Artery - Aorta
\item Artery - Coronary
\item Artery - Tibial
\item Bladder
\item Brain - Amygdala
\item Brain - Anterior cingulate cortex (BA24)
\item Brain - Caudate (basal ganglia)
\item Brain - Cerebellar Hemisphere
\item Brain - Cerebellum
\item Brain - Cortex
\item Brain - Frontal Cortex (BA9)
\item Brain - Hippocampus
\item Brain - Hypothalamus
\item Brain - Nucleus accumbens (basal ganglia)
\item Brain - Putamen (basal ganglia)
\item Brain - Spinal cord (cervical c-1)
\item Brain - Substantia nigra
\item Breast - Mammary Tissue
\item Cells - EBV-transformed lymphocytes
\item Cells - Transformed fibroblasts
\item Cervix - Ectocervix
\item Cervix - Endocervix
\item Colon - Sigmoid
\item Colon - Transverse
\item Esophagus - Gastroesophageal Junction
\item Esophagus - Mucosa
\item Esophagus - Muscularis
\item Fallopian Tube
\item Heart - Atrial Appendage
\item Heart - Left Ventricle
\item Kidney - Cortex
\item Liver
\item Lung
\item Minor Salivary Gland
\item Muscle - Skeletal
\item Nerve - Tibial
\item Ovary
\item Pancreas
\item Pituitary
\item Prostate
\item Skin - Not Sun Exposed (Suprapubic)
\item Skin - Sun Exposed (Lower leg)
\item Small Intestine - Terminal Ileum
\item Spleen
\item Stomach
\item Testis
\item Thyroid
\item Uterus
\item Vagina
\item Whole Blood
}}

\item{datasetId}{A character string. Options: "gtex_v8" (default), "gtex_v7".}
}
\value{

}
\description{
This function fetch queried genes' expression profiles in a tissue at the sample level.
}
\examples{
 expProfiles <- GTExquery_exp(c("tp53","naDK","SDF4"), "symbol", "Liver", "gtex_v8")
 expProfiles <- GTExquery_exp(c("tp53","naDK","SDF4"), "symbol", "Artery - Coronary", "gtex_v8")
 expProfiles <- GTExquery_exp(c("tp53","naDK","SDF4"), "symbol", "Artery - Coronary", "gtex_v7")
}
