% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{GTExquery_gene}
\alias{GTExquery_gene}
\title{Query gene info through API.}
\usage{
GTExquery_gene(
  genes = "",
  geneType = "symbol",
  gencodeVersion = "v26",
  recordPerChunk = 150
)
}
\arguments{
\item{gencodeVersion}{"v26" or "v19"}

\item{recordPerChunk}{A integer value. Defaulut: 150}
}
\value{
A data.frame
}
\description{
Query gene info through API.
}
\examples{
 miRNA <- apiRef_genes(genes="miRNA", geneType="geneCategory", "v26","GRCh38/hg38" )
 a <- GTExquery_gene(miRNA$geneSymbol,"v26", 150)
}
