% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{GTExquery_tissue}
\alias{GTExquery_tissue}
\title{Fetch all tissue information for a dataset.}
\usage{
GTExquery_tissue(tissueName = "", datasetId = "gtex_v8")
}
\arguments{
\item{tissueName}{Tissue name, tissue ID or tissue site name. Default return all tissues' information. Can be choonse from \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{datasetId}{gtex_v8 or gtex_v7. Default: gtex_v8}
}
\value{
A data.table
}
\description{
Information includes tissue IDs, number of RNA-Seq samples, number of RNA-Seq samples with genotype, number of expressed genes, number of eGenes. Also includes tissueSiteDetail ID, name, abbreviation, uberon ID, and standard tissue colors. TissueSiteDetails are grouped by TissueSites. By default, this service reports from the latest GTEx release.
}
\examples{
\donttest{
  tissueAll <- GTExquery_tissue(datasetId="gtex_v8")
  BrainInfo <- GTExquery_tissue("Brain", datasetId="gtex_v8")
}
}
