% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{GTExquery_varPos}
\alias{GTExquery_varPos}
\title{Query variant information by position.}
\usage{
GTExquery_varPos(
  chrom = "",
  pos = numeric(0),
  datasetId = "gtex_v8",
  recordPerChunk = 200
)
}
\arguments{
\item{chrom}{A character string.}

\item{pos}{An integer array.}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}

\item{recordPerChunk}{A integer value (1-200). number of records fetched per request (default: 200).}
}
\value{
A data.table.
}
\description{
Query variant information by position.
}
\examples{
\donttest{
 GTExquery_varPos(chrom="chr1", pos=c(1102708,1105739),"gtex_v8")
 GTExquery_varPos(chrom="1", pos=c(1038088,1041119),"gtex_v7")
 GTExquery_varPos("1", c(1246438, 1211944, 1148100),"gtex_v7")
}
}
