% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_eqtlGWAS}
\alias{GTExvisual_eqtlGWAS}
\title{significance analysis of GTEX and eQTL  NO.}
\usage{
GTExvisual_eqtlGWAS(
  gwasDF,
  traitGene = "",
  traitGeneType = "geneSymbol",
  tissueSiteDetail = "",
  study_id = "gtex_v8"
)
}
\arguments{
\item{gwasDF}{A data.frame containing GWAS summary info. At least four columns are expected, including "snpId", "chrom", "pos" and "pValue".}

\item{traitGene}{eqtl trait. If this parameter is not specified, it is the same as the "queryTerm".}

\item{traitGeneType}{eqtl type. If this parameter is not specified, it is the same as the "queryType".}

\item{tissueSiteDetail}{tissue}

\item{study_id}{A character string. Default: "gtex_v8".}
}
\value{
A plot.
}
\description{
significance analysis of GTEX and eQTL  NO.
}
\examples{
\donttest{


}
}
