% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_eqtlTrait}
\alias{GTExvisual_eqtlTrait}
\title{}
\usage{
GTExvisual_eqtlTrait(
  gene = "",
  geneType = "geneSymbol",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\value{

}
\description{

}
\examples{
\donttest{
 GTExvisual_eqtlTrait("ENSG00000112137.12", "gencodeId",tissueSiteDetail = "Adipose - Subcutaneous", datasetId="gtex_v7")
}
}
