% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_geneExpviolin}
\alias{GTExvisual_geneExpviolin}
\title{Plot distribution of gene expression among multiple tissues.}
\usage{
GTExvisual_geneExpviolin(
  gene = "",
  geneType = "geneSymbol",
  datasetId = "gtex_v8",
  toTissueSite = TRUE
)
}
\arguments{
\item{geneType}{"geneSymbol" or "gencodeId".}

\item{datasetId}{"gtex_v8" or "gtex_v7".}

\item{toTissueSite}{TRUE or FALSE, display all subtissues or tissue Site. Default: TURE.}

\item{genes}{A characer vector.}
}
\value{
A data.table and a plot.
}
\description{
Plot distribution of gene expression among multiple tissues.
}
\examples{
\donttest{
  gene = c("ENSG00000069812.11", "ENSG00000141510.16")
  gene="TP53"
  gene="HES3"
  geneType="gencodeId"
  a <- GTExvisual_geneExpviolin("ENSG00000069812.11",toTissueSite=TRUE)
}
}
