% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_locusZoom}
\alias{GTExvisual_locusZoom}
\title{LocusZoom plot}
\usage{
GTExvisual_locusZoom(
  DF,
  highlightSnp = "",
  population = "EUR",
  posRange = "",
  token = "9246d2db7917",
  windowSize = 5e+05,
  genomeVersion = "grch38",
  snpLD = NA
)
}
\arguments{
\item{DF}{A data.frame or a data.table object. Four columns are required: "snps", a character list, using an rsID or chromosome coordinate (e.g. "chr7:24966446"); chromosome, chr1-chr22; Genome position; P-value.}

\item{highlightSnp}{Default is the snp that with lowest p-value.}

\item{population}{Supported population is consistent with the LDlink, which can be listed using function "LDlinkR::list_pop()"}

\item{posRange}{visualized genome region of interest. Default is the region that covers all snps.}

\item{token}{LDlink provided user token, default = NULL, register for token at https://ldlink.nci.nih.gov/?tab=apiaccess}

\item{windowSize}{Window around the highlighted snp for querying linkage disequilibrium information. Default:500000}

\item{genome}{"grch38"(default) or "grch37".}
}
\value{
A data.table object and plot.
}
\description{
LocusZoom plot
}
\examples{
\donttest{
 # For GWAS:
 gwasFile <- tempfile(pattern = "file")
 gwasURL <- "https://raw.githubusercontent.com/dingruofan/exampleData/master/gwas/AD/gwasChr6Sub1.txt"
 utils::download.file(gwasURL, destfile=gwasFile)
 gwasDF <- data.table::fread(gwasFile, sep="\t", header=TRUE)
 gwasDF <- gwasDF[,.(rsid, chr, position,P)]
 GTExvisual_locusZoom(gwasDF)
 GTExvisual_locusZoom(gwasDF, posRange="chr6:3e7-7e7", population ="AFR", windowSize=200000)
 GTExvisual_locusZoom(gwasDF, posRange="chr6:3e7-7e7", population ="AFR", windowSize=500000, highlightSnp="rs9271165")

 # For eQTL:
 eqtlAsso <- GTExdownload_assoAll("RP11-385F7.1", tissueSiteDetail = "Brain - Cortex", withB37VariantId=FALSE)
 eqtlAsso[,c("chrom","pos")]<-rbindlist(lapply(eqtlAsso$variantId, function(x){ a=stringr::str_split(x,"_")[[1]];return(data.table(chrom=a[1], pos=a[2])) }))
 GTExvisual_locusZoom( eqtlAsso[,.(snpId, chrom, pos, pValue)], population="EUR",
                      posRange="chr6:46488310-48376712", genomeVersion="grch38" )
}
}
