% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dealWithGencode.R
\name{gtfSubsGene}
\alias{gtfSubsGene}
\title{Extract gene attributes of interest}
\usage{
gtfSubsGene(
  gtf_attributes,
  att_of_interest = c("gene_id", "gene_type", "gene_name")
)
}
\arguments{
\item{gtf_attributes}{like: c("gene_id", "gene_type", "gene_name")}
}
\value{
specificed attributes
}
\description{
as a funciton of lapply
}
\examples{
 \dontrun{
  # extract gene info:
  gencodeENSG <- data.table::rbindlist(lapply(gencodeAnnoGene$attributes, gtfSubsGene, att_of_interest= c("gene_id", "gene_type", "gene_name")))
  # extract transcript info:
  gencodeENSG <- data.table::rbindlist(lapply(gencodeAnnoGene$attributes, gtfSubsGene, att_of_interest= c("gene_id","transcript_id", "gene_type", "gene_name")))
 }
}
