% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dealWithGencode.R
\name{gtfSubsGeneInfo}
\alias{gtfSubsGeneInfo}
\title{Title download gtf files and extract gene info from attribute column:}
\usage{
gtfSubsGeneInfo(gencodeVersion = "v26")
}
\arguments{
\item{gencodeVersion}{}
}
\value{

}
\description{
Title download gtf files and extract gene info from attribute column:
}
\examples{
 gtfSubsGeneInfo("v26")
}
