% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieveLD_LDmatrix}
\alias{retrieveLD_LDmatrix}
\title{fetch matrix of pairwise linkage disequilibrium statistics.}
\usage{
retrieveLD_LDmatrix(snps, pop = "CEU", r2d = "r2", token = NULL, file = FALSE)
}
\arguments{
\item{snps}{list of between 2 - 1,000 variants, using an rsID or chromosome coordinate (e.g. "chr7:24966446")}

\item{pop}{a 1000 Genomes Project population, (e.g. YRI or CEU), multiple allowed, default = "CEU". All supported can be listed with LDlinkR::list_pop()}

\item{r2d}{r2d, either "r2" for LD R2 or "d" for LD D', default = "r2"}

\item{token}{LDlink provided user token, default = NULL, register for token at https://ldlink.nci.nih.gov/?tab=apiaccess}

\item{file}{Optional character string naming a path and file for saving results. If file = FALSE, no file will be generated, default = FALSE.}
}
\value{
a data frame
}
\description{
This fucntion is rebuilt from LDlink package.
}
\examples{
\donttest{
  ldMat <- retrieveLD_LDmatrix(c("rs12202891", "rs10807323",
                     "rs9381401", "rs2026458", "rs150503442"),
                      pop="CEU", token="9246d2db7917")
}
}
