% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExbiolinks.R
\docType{data}
\name{sampleNum}
\alias{sampleNum}
\title{samples used in GTEx eQTL analysis.}
\format{
A data frame with 49 rows and 2 variables:
\describe{
\item{tissueSiteDetailId}{character string, tissue name}
\item{sampleNum}{integer}
}
}
\source{
\url{https://gtexportal.org/home/}
}
\description{
A dataset containing the 49 tissues' name and corresponding sample number.
}
\keyword{internal}
