% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{tissueSiteDetailGTExv7}
\alias{tissueSiteDetailGTExv7}
\title{Tissue name and tissue id mapping of GTEx V7.}
\format{
A data frame with 53 rows and 2 variables:
}
\source{
\url{https://gtexportal.org/home/}
}
\description{
A dataset containing the 53 tissues' name and corresponding ID of GTEx V7.
}
\keyword{internal}
