% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{tissueSiteDetailGTExv8}
\alias{tissueSiteDetailGTExv8}
\title{Tissue name and tissue id mapping of GTEx V8.}
\format{
A data frame with 54 rows and 2 variables
}
\source{
\url{https://gtexportal.org/home/}
}
\description{
A dataset containing the 54 tissues' name and corresponding ID of GTEx V8.
}
\keyword{internal}
