% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTL_export}
\alias{xQTL_export}
\title{export expression object to specified format}
\usage{
xQTL_export(exp_object, out_format = "to_clusterP")
}
\arguments{
\item{exp_object}{expression object derived from \code{xQTLdownload_exp}}

\item{out_format}{"to_clusterP", "to_wgcna" and to "to_deseq"}
}
\value{
A data.frame/data.table object
}
\description{
export expression object to specified format
}
\examples{
\donttest{
expProfiles <- xQTLdownload_exp(c("tp53","naDK","SDF4"),
                               tissueSiteDetail="Artery - Coronary",
                               pathologyNotesCategories=TRUE, toSummarizedExperiment = TRUE)
}
}
