% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_anno}
\alias{xQTLanalyze_anno}
\title{annotate variants with genome positon}
\usage{
xQTLanalyze_anno(snpInfo = "", genomeVersion = "hg38")
}
\arguments{
\item{snpInfo}{A data.table/data.frame with two or three columns: chromosome and position.}

\item{genomeVersion}{"hg38" (default) or "hg19". Note: hg19 will be converted to hg38 automatically.}
}
\value{
A data.table object of variants' genomics distribution
}
\description{
annotate variants with genome positon
}
\examples{
\donttest{
url1 <- "http://github.com/dingruofan/exampleData/raw/master/gwas/gwasSub.txt.gz"
snpInfo <- fread(url1, sep="\t")
snpHits <- xQTLanalyze_anno(snpInfo)
}
}
