% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_calLambda}
\alias{xQTLanalyze_calLambda}
\title{calculate genomic control inflation factor}
\usage{
xQTLanalyze_calLambda(summaryDT)
}
\arguments{
\item{summaryDT}{A data.frame containing one or two columns: p-value (required) and group (optional)}
}
\value{
A data.table object
}
\description{
calculate genomic control inflation factor
}
\examples{
\donttest{
url1 <- "https://raw.githubusercontent.com/dingruofan/exampleData/master/eqtl/MMP7_qtlDF.txt"
qtl <- data.table::fread(url1, sep="\t")

# calculate lambda value with all variants
xQTLanalyze_calLambda(qtl[,.(pValue)])

# calculate lambda value for each group:
qtl$groups <- sample(c(0,1),size = nrow(qtl), replace = TRUE)
xQTLanalyze_calLambda(qtl[,.(pValue, groups)])
}
}
