% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_coloc_local}
\alias{xQTLanalyze_coloc_local}
\title{conduct colocalization analysis with your own data}
\usage{
xQTLanalyze_coloc_local(
  gwasDF,
  qtlDF,
  mafThreshold = 0.01,
  gwasSampleNum = 50000,
  qtlSampleNum = 10000,
  method = "coloc",
  bb.alg = FALSE
)
}
\arguments{
\item{gwasDF}{data.frame or data.table, required cols: rsid, chrom, position, pValue, maf, beta, se}

\item{qtlDF}{data.frame or data.table, required cols:  rsid, chrom, position, pValue, maf, beta, se}

\item{mafThreshold}{0.01}

\item{gwasSampleNum}{50000}

\item{qtlSampleNum}{10000}

\item{method}{"coloc", "hyprcoloc", and "Both"}

\item{bb.alg}{TRUE}
}
\value{
A list
}
\description{
conduct colocalization analysis with your own data
}
