% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_enrich}
\alias{xQTLanalyze_enrich}
\title{variant enrichment analysis}
\usage{
xQTLanalyze_enrich(
  snpInfo = "",
  genomeVersion = "hg38",
  enrichElement = "Promoter",
  distLimit = 1e+06
)
}
\arguments{
\item{snpInfo}{A data.table/data.frame with two or three columns: chromosome and position.}

\item{genomeVersion}{"hg38" (default) or "hg19". Note: hg19 will be converted to hg38 automatically.}

\item{enrichElement}{"Promoter", "Enhaner" or "TF".}

\item{distLimit}{Defaults: 1e6}
}
\value{
A data.table object
}
\description{
variant enrichment analysis
}
\examples{
\donttest{
url1 <- "http://github.com/dingruofan/exampleData/raw/master/gwas/gwasSub.txt.gz"
snpInfo <- fread(url1, sep="\t")
enrichHits <- xQTLanalyze_enrich(snpInfo)
}
}
