% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_getSentinelSnp}
\alias{xQTLanalyze_getSentinelSnp}
\title{xQTLanalyze_getSentinelSnp}
\usage{
xQTLanalyze_getSentinelSnp(
  gwasDF,
  pValueThreshold = 5e-08,
  centerRange = 10000,
  mafThreshold = 0.01,
  genomeVersion = "grch38",
  grch37To38 = FALSE
)
}
\arguments{
\item{gwasDF}{GWAS data.frame}

\item{pValueThreshold}{Cutoff of gwas p-value. Default: 5e-8}

\item{centerRange}{SNP-to-SNP distance. Default:1e4}

\item{mafThreshold}{Cutoff of maf to remove rare variants.}

\item{genomeVersion}{Genome version of input file. "grch37" or "grch38" (default).}

\item{grch37To38}{TRUE or FALSE, we recommend converting grch37 to grch38, or using a input file of grch38 directly. Package \code{rtracklayer} is required.}
}
\value{
A data.table object.
}
\description{
detect sentinel SNPs in a given GWAS dataset.
Return sentinel snps whose pValue < 5e-8(default) and SNP-to-SNP distance > 1e4 bp.
}
\examples{
\donttest{
   gwasFile <- tempfile(pattern = "file")
   gwasURL <- paste0("https://raw.githubusercontent.com/dingruofan/",
                     "exampleData/master/gwas/AD/GLGC_AD_chr1_6_Sub3.txt")
   utils::download.file(gwasURL, destfile=gwasFile)
   gwasDF <- data.table::fread(gwasFile, sep="\t")
   gwasDF <- gwasDF[, .(rsid, chr, position, P, maf)]
   sentinelSnpDF <- xQTLanalyze_getSentinelSnp(gwasDF)
}
}
