% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_qtlSpecificity}
\alias{xQTLanalyze_qtlSpecificity}
\title{eQTL pvalues in different LD bins.}
\usage{
xQTLanalyze_qtlSpecificity(
  gene = "",
  geneType = "auto",
  variantName = "",
  variantType = "auto",
  study = "",
  population = "EUR"
)
}
\arguments{
\item{gene}{gene A gene symbol or a gencode id (versioned).}

\item{geneType}{geneType A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{study}{Studies can be listed using "ebi_study_tissues"}

\item{population}{(string) One of the 5 popuations from 1000 Genomes: 'AFR', 'AMR', 'EAS', 'EUR', and 'SAS'.}
}
\value{
a ggplot2 object
}
\description{
eQTL pvalues in different LD bins.
}
\examples{
p <- xQTLanalyze_qtlSpecificity(gene="MMP7", variantName="rs11568818")
}
