% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_TSExp}
\alias{xQTLanalyze_TSExp}
\title{xQTLanalyze_TSExp}
\usage{
xQTLanalyze_TSExp(
  genes,
  geneType = "auto",
  method = "SPM",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{genes}{A charater vector or a string of gene symbol, gencode id (versioned), or a charater string of gene type.}

\item{geneType}{A character string. "auto"(default), "geneSymbol", "gencodeId" or "geneCategory".}

\item{method}{"SPM" or "entropy"}

\item{datasetId}{"gtex_v8" or "gtex_v7".}
}
\value{
A data.table
}
\description{
perform tissue-specific expression analysis.
}
\examples{
TSgene <- xQTLanalyze_TSExp(extractGeneInfo(gencodeGeneInfoAllGranges)$gencodeId[1:20])
# xQTLvisual_geneExpTissues( TSgene[order(-DPM)][1,]$geneSymbol )
}
