% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{package}
\name{xQTLbiolinks}
\alias{xQTLbiolinks}
\title{xQTLbiolinks package}
\description{
the functions provided in this \pkg{xQTLbiolinks} enable users to access molecular QTLs (eQTLs and sQTLs) and gene expressions data filtered by tissue, gene, variant or dataset.
}
