% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_eqtl}
\alias{xQTLdownload_eqtl}
\title{xQTLdownload_eqtl}
\usage{
xQTLdownload_eqtl(
  variantName = "",
  gene = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = "",
  datasetId = "gtex_v8",
  recordPerChunk = 100
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{gene}{A gene symbol or a gencode id (versioned). Can not be null.}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{A character string. tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}

\item{recordPerChunk}{A integer value (1-500). number of records fetched per request (default: 100).}
}
\value{
A data.table
}
\description{
download significant or unsignificant eQTL data of a tissue or across all tissues.
can be quried with a gene/variant-gene pair.
}
\examples{
\donttest{

 # Download eQTL info for a gene:
 eqtlInfo <- xQTLdownload_eqtl(gene="TP53")
 xQTLdownload_eqtl(gene="ATAD3B", datasetId="gtex_v7")

 # Unversioned gencode ID in GTEx V7:
 eqtl_v7 <- xQTLdownload_eqtl(gene="ENSG00000141510",
                              datasetId="gtex_v7")
 # Unversioned gencode ID in GTEx V8:
 eqtl_v8 <- xQTLdownload_eqtl(gene="ENSG00000141510",
                              datasetId="gtex_v8")

 # In a specific tissue:
 xQTLdownload_eqtl(gene="ENSG00000141510.11", geneType="gencodeId",
                      datasetId="gtex_v7", tissueSiteDetail="Thyroid" )

 # Download eQTL info for a variant-gene pair:
 xQTLdownload_eqtl(variantName="rs1641513",gene="TP53", datasetId="gtex_v8")
 xQTLdownload_eqtl(variantName="rs11657498",gene="TP53",
                      datasetId="gtex_v7")
 xQTLdownload_eqtlSig(variantName="chr1_1667948_A_G_b38", gene="SLC35E2B",
                      tissueSiteDetail="Kidney - Cortex")
 xQTLdownload_eqtl(variantName="17_7492388_G_A_b37",gene="TP53",
                      tissueSiteDetail="Uterus",  datasetId="gtex_v7")
}
}
