% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{xQTLdownload_eqtlAllAsso}
\alias{xQTLdownload_eqtlAllAsso}
\title{This function return the gene-variant association for any given pair of gene and variant, which may or may not be significant}
\usage{
xQTLdownload_eqtlAllAsso(
  gene = "",
  geneType = "auto",
  tissueSiteDetail = "",
  recordPerChunk = 250,
  study = "gtex_v8",
  withB37VariantId = TRUE
)
}
\arguments{
\item{gene}{gene A gene symbol or a gencode id (versioned).}

\item{geneType}{geneType A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{tissueSiteDetail A character string. tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{recordPerChunk}{A integer value (1-200). number of records fetched per request (default: 200).}

\item{study}{"gtex_v8" only GTEx v8 is supported.}

\item{withB37VariantId}{Whether to return the genome location(GTEx v7) of variants. Default: TRUE.}

\item{genes}{GENCODE ID (versioned or unversioned).}

\item{variants}{GTEx variant ID or dbSNP ID. Note: if variant id are provided, genome version of hg38 are required.}
}
\value{
A data.table object

a data.table
}
\description{
Only GTEx V8 are supported.

download all tested variant-gene associations.
source of all eQTL associations is EBI eQTL category.
}
\examples{
\donttest{
   geneList = c("ENSG00000228463","ENSG00000228463.9", "ENSG00000065613.13")
   variants = c("rs201327123","chr1_14677_G_A_b38", "chr11_66561248_T_C_b38")
   xQTLdownload_eqtlPost(geneList, variants, tissueSiteDetail="Colon - Sigmoid")
}
\donttest{
  geneAsso <- xQTLdownload_eqtlAllAsso("ATP11B", tissueSiteDetail="Muscle - Skeletal", withB37VariantId=FALSE)
}
}
