% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{xQTLdownload_eqtlAllPost}
\alias{xQTLdownload_eqtlAllPost}
\title{This function return the gene-variant association for any given pair of gene and variant, which may or may not be significant}
\usage{
xQTLdownload_eqtlAllPost(
  geneList,
  variantlist,
  tissueSiteDetail = "",
  recordPerChunk = 50
)
}
\arguments{
\item{tissueSiteDetail}{tissue site detail.}

\item{recordPerChunk}{A integer value (1-200). number of records fetched per request (default: 50).}

\item{genes}{GENCODE ID (versioned or unversioned).}

\item{variants}{GTEx variant ID or dbSNP ID. Note: if variant id are provided, genome version of hg38 are required.}
}
\value{
A data.table object
}
\description{
Only GTEx V8 are supported.
}
\examples{
\donttest{
   geneList = c("ENSG00000228463","ENSG00000228463.9", "ENSG00000065613.13")
   variants = c("rs201327123","chr1_14677_G_A_b38", "chr11_66561248_T_C_b38")
   xQTLdownload_eqtlAllPost(geneList, variants, tissueSiteDetail="Colon - Sigmoid")
}
}
