% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_exp}
\alias{xQTLdownload_exp}
\title{Download normalized gene expression at the sample level in a specified tissue.}
\usage{
xQTLdownload_exp(
  genes = "",
  geneType = "auto",
  tissueSiteDetail = "Liver",
  datasetId = "gtex_v8",
  toSummarizedExperiment = TRUE,
  recordPerChunk = 80,
  pathologyNotesCategories = FALSE
)
}
\arguments{
\item{genes}{(character string or a character vector) gene symbols or gencode ids (versioned or unversioned are both supported).}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{(character) details of tissues in GTEx can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{datasetId}{(character) options: "gtex_v8" (default), "gtex_v7".}

\item{toSummarizedExperiment}{a logical value indicating whether to return a data.frame or a summarizedExperiment object. Default: TRUE, return a toSummarizedExperiment object.}

\item{recordPerChunk}{(integer) number of records fetched per request (default: 80).}

\item{pathologyNotesCategories}{a logical value indicating whether to return pathologyNotes. Default: FALSE, the pathologyNotes is ignored.}
}
\value{
return a SummarizedExperiment or a data.table object harbing gene expression profiles and samples' information.
}
\description{

}
\examples{
# Download gene expression with a genecode ID:
expProfiles <- xQTLdownload_exp("ENSG00000210195.2", tissueSiteDetail="Liver")

# extract expression profile from SummarizedExperiment object:
expDT <- SummarizedExperiment::assay(expProfiles)

# extract samples' detail from SummarizedExperiment object:
sampleDT <- SummarizedExperiment::colData(expProfiles)

# Download gene expression profiles of multiple genes:
expProfiles <- xQTLdownload_exp(c("tp53","naDK","SDF4"),
                                tissueSiteDetail="Artery - Coronary",
                                pathologyNotesCategories=TRUE,
                                toSummarizedExperiment=FALSE)
expProfiles <- xQTLdownload_exp(c("tp53","naDK","SDF4"),
                                tissueSiteDetail="Artery - Coronary",
                                datasetId="gtex_v7")

# Get proteing-coding genes' expression in Lung:
# prot <- xQTLquery_gene(genes="protein coding")
# proTexp <- xQTLdownload_exp(proT$geneSymbol, tissueSiteDetail="Lung",
#                             toSummarizedExperiment=FALSE)
}
