% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_hqtl}
\alias{xQTLdownload_hqtl}
\title{Download summary statistics of H3K4me1 and H3K27ac histone QTL (hQTL) of a specified location}
\usage{
xQTLdownload_hqtl(
  phenotype_id = "9:99773935-99776816",
  histone_type = "H3K27AC",
  cell_type = "monocyte",
  hqtlmeta = NULL
)
}
\arguments{
\item{phenotype_id}{phenotype_id that formatted with genome location, like: 9-99773935-99776816, can be obtained using \code{xQTLdownload_hqtlmeta}}

\item{histone_type}{(string) One of the histone types: "H3K27AC" or "H3K4ME1".}

\item{cell_type}{(string) One of the cell types: "monocyte", "neutrophil" or "T cell".}

\item{hqtlmeta}{A data.table object obtained via \code{xQTLdownload_hqtlmeta}.}
}
\value{
A data.table object
}
\description{
Download summary statistics of H3K4me1 and H3K27ac histone QTL (hQTL) of a specified location
}
\examples{
 hQTL_dt <- xQTLdownload_hqtl(phenotype_id="10:10458128-10465096", histone_type="H3K4ME1", cell_type="T cell")
}
