% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_hqtlmeta}
\alias{xQTLdownload_hqtlmeta}
\title{download meta data of H3K4me1 and H3K27ac histone QTL (hQTL)}
\usage{
xQTLdownload_hqtlmeta(histone_type = "H3K27AC", cell_type = "monocyte")
}
\arguments{
\item{histone_type}{(string) One of the histone types: "H3K27AC" or "H3K4ME1".}

\item{cell_type}{(string) One of the cell types: "monocyte", "neutrophil" or "T cell".}
}
\value{
a data.table object includng all CpG ID
}
\description{
download meta data of H3K4me1 and H3K27ac histone QTL (hQTL)
}
\examples{
hqtlmeta <- xQTLdownload_hqtlmeta(histone_type="H3K4ME1", cell_type="T cell")
}
