% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_ld}
\alias{xQTLdownload_ld}
\title{Download linkage disequilibrium data of the variants associated with this gene.}
\usage{
xQTLdownload_ld(
  gene = "",
  geneType = "geneSymbol",
  datasetId = "gtex_v8",
  recordPerChunk = 300
)
}
\arguments{
\item{gene}{(character) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{geneType}{(character) "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{(character) options: "gtex_v8" (default), "gtex_v7".}

\item{recordPerChunk}{(integer) number of records fetched per request (default: 100).}
}
\value{
A data.table object.
}
\description{
Download linkage disequilibrium data of the variants associated with this gene.
}
\examples{
xQTLdownload_ld("TP53" )
xQTLdownload_ld("TP53", datasetId="gtex_v7")
xQTLdownload_ld(gene="ENSG00000008128.22")
}
