% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{xQTLdownload_sqtlSig}
\alias{xQTLdownload_sqtlSig}
\title{Fetch all eQTL associations from EBI eQTL category.}
\usage{
xQTLdownload_sqtlSig(
  variantName = "",
  gene = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = ""
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{gene}{A gene symbol or a gencode id (versioned).}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{A character string. tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{refSeq}{A character string.}

\item{rsid}{dbSNP ID.}

\item{variantId}{A character string. like "chr17_5294580_CT_C_b38".}

\item{pThreshold}{Float. set a threshold for p-value to filter the result. Default: 1, aQTL associations that have a p-value<1 will be returned.}
}
\value{


A data.table
}
\description{
download significant sQTL associations of a tissue or across all tissues.
Only GTEx v8 is supported.
}
\examples{
\donttest{
  xQTLdownload_aQTLAllAsso( gene="TP53", tissueSiteDetail = "Adipose - Visceral (Omentum)" )
  xQTLdownload_aQTLAllAsso( refSeq = "NM_001291581.2", tissueSiteDetail = "Adipose - Visceral (Omentum)" )
  xQTLdownload_aQTLAllAsso( rsid = "rs3026133", tissueSiteDetail = "Adipose - Visceral (Omentum)", pThreshold=1e-1 )
  xQTLdownload_aQTLAllAsso( variantId = "chr17_5289661_A_G_b38", tissueSiteDetail = "Adipose - Visceral (Omentum)" )
}
\donttest{
 # Download sQTL info for a variant:
 xQTLdownload_sqtlSig(variantName="rs201327123")
 xQTLdownload_sqtlSig(variantName="chr1_14677_G_A_b38")
 xQTLdownload_sqtlSig(variantName="chr1_14677_G_A_b38",
                      tissueSiteDetail="Whole Blood")

 # Download sQTL association according to all tissues with genome location:
 xQTLquery_varPos(chrom="chr1", pos=c(1102708),"gtex_v8")
 xQTLdownload_sqtlSig(variantName=varInfo$snpId, variantType="snpId")

 # Download sQTL info for a gene:
 xQTLdownload_sqtlSig(gene="ATAD3B")
 xQTLdownload_sqtlSig(gene="ENSG00000141510.16")
 xQTLdownload_sqtlSig(gene="ENSG00000141510.16",
                      tissueSiteDetail="Lung" )

 # Download eQTL info for a variant-gene pair:
 xQTLdownload_sqtlSig(variantName="rs546057177", gene="TP53")
 xQTLdownload_sqtlSig(variantName="chr17_7465085_A_G_b38",
                      gene="TP53", tissueSiteDetail="Lung")
}
}
