% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_xqtlAllAsso}
\alias{xQTLdownload_xqtlAllAsso}
\title{Download summary statistics of xQTL of a specified gene in GTEx v8, default:3'aQTL}
\usage{
xQTLdownload_xqtlAllAsso(
  genes = "",
  geneType = "",
  tissue = "",
  mRNA_refseq = "",
  mRNA_gene_DF = NULL,
  type = "3'aQTL"
)
}
\arguments{
\item{genes}{(character) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissue}{(character) details of tissues in GTEx can be listed using \code{tissueSiteDetailGTExv8}}

\item{mRNA_refseq}{(character) If provided, only the 3'aQTL of mRNA will be downloaded}

\item{mRNA_gene_DF}{(data.frame) If provided, mRNA-gencode mapping relationship will be loaded from this data.frame.}

\item{type}{3'aQTL(default)}
}
\value{
A data.table object of xQTL dataset.
}
\description{
Download summary statistics of xQTL of a specified gene in GTEx v8, default:3'aQTL
}
\examples{
\donttest{
aQTL_DT <- xQTLdownload_xqtlAllAsso(genes=c("MMP7"), tissue="Lung")
}
}
