% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{xQTLquery_gene}
\alias{xQTLquery_gene}
\title{xQTLquery_gene.}
\usage{
xQTLquery_gene(
  genes = "",
  geneType = "geneSymbol",
  gencodeVersion = "v26",
  recordPerChunk = 150
)
}
\arguments{
\item{genes}{A charater vector or a string of gene symbol, gencode id (versioned), or a charater string of gene type.
\itemize{
\item \strong{gene symbol (Default)}.

A character string or a character vector (case ignored). like: "tp53","naDK","SDF4".
\item \strong{gencode/ensemble id} (versioned or unversioned).

A character string or a character vector (case ignored). like: "ENSG00000210195.2","ENSG00000078808".
\item \strong{gene classification}.

"genes" must be chosen from following list when "geneType" is "geneCategory".
\itemize{
\item protein coding
\item antisense
\item lincRNA
\item unprocessed pseudogene
\item miRNA
\item transcribed unprocessed pseudogene
\item snRNA
\item processed pseudogene
\item processed transcript
\item TEC
\item transcribed unitary pseudogene
\item transcribed processed pseudogene
\item sense intronic
\item misc RNA
\item snoRNA
\item scaRNA
\item rRNA
\item unitary pseudogene
\item 3prime overlapping ncRNA
\item polymorphic pseudogene
\item bidirectional promoter lncRNA
\item sense overlapping
\item pseudogene
\item IG V pseudogene
\item scRNA
\item IG C gene
\item IG J gene
\item IG V gene
\item sRNA
\item ribozyme
\item vaultRNA
\item non coding
\item TR J gene
\item TR C gene
\item TR V gene
\item TR V pseudogene
\item TR D gene
\item IG C pseudogene
\item macro lncRNA
\item TR J pseudogene
\item IG D gene
\item IG J pseudogene
\item IG pseudogene
\item Mt tRNA
\item Mt rRNA
}
}}

\item{geneType}{A character string. "geneSymbol"(default), "gencodeId" or "geneCategory".}

\item{gencodeVersion}{"v26" or "v19"}

\item{recordPerChunk}{A integer value. Defaulut: 150}
}
\value{
A data.table of queried gene information. including following columns:
\itemize{
\item \strong{genes.} Input genes
\item \strong{geneSymbol.} Gene symbol.
\item \strong{gencodeId.} Gencode/ensemble id (versioned).
\item \strong{entrezGeneId.} Entrez gene ID.
\item \strong{geneType.} Gene type.
\item \strong{chromosome.} Note: "chr" is added in gencode v26,
\item \strong{start.}
\item \strong{end.}
\item \strong{strand.}
\item \strong{tss.} Transcriptional start site.
\item \strong{gencodeVersion.} Gencode Version.
\item \strong{genomeBuild.} Genome version.
\item \strong{description.}
}
}
\description{
users can use this function to query basic information (including gene name, symbol, position and description, etc ) of genes.
}
\examples{
\donttest{
 # get all protein coding genes and description:
 protein_coding <- xQTLquery_gene(genes="protein coding", geneType="geneCategory", "v26" )
 # get all miRNA and description:
 miRNA <- xQTLquery_gene(genes="miRNA", geneType="geneCategory", "v19")

 # hg38 test:
 geneInfo <- xQTLquery_gene("TP53", "geneSymbol", "v26")
 geneInfo <- xQTLquery_gene(c("tp53","naDK","SDF4"), "geneSymbol", "v26")
 geneInfo <- xQTLquery_gene(c("ENSG00000210195.2","ENSG00000078808"),
                              geneType="gencodeId", "v26")
 # hg19 test:
 geneInfo <- xQTLquery_gene(c("TP53","naDK"), "geneSymbol", "v19")
 geneInfo <- xQTLquery_gene(c("ENSG00000141510.11","ENSG00000008130.11"), "gencodeId", "v19")
 }
}
