% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{xQTLquery_geneAll}
\alias{xQTLquery_geneAll}
\title{xQTLquery_geneAll}
\usage{
xQTLquery_geneAll(gencodeVersion = "v26", recordPerChunk = 2000)
}
\arguments{
\item{gencodeVersion}{A character string. "v26" or "v19".}

\item{recordPerChunk}{A integer value. From 1 to 2000.}
}
\value{
A data.table with all genes' information.
}
\description{
fetch basic information of all genes.
}
\examples{
# Fetch all genes. This will take several minutes:
# allGenes <- xQTLquery_geneAll()
}
