% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{xQTLquery_sampleByTissue}
\alias{xQTLquery_sampleByTissue}
\title{Query sample's details with tissue name}
\usage{
xQTLquery_sampleByTissue(
  tissueSiteDetail = "Liver",
  dataType = "RNASEQ",
  datasetId = "gtex_v8",
  recordPerChunk = 200,
  pathologyNotesCategories = FALSE
)
}
\arguments{
\item{tissueSiteDetail}{(character) details of tissues in GTEx can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{dataType}{A character string. Options: "RNASEQ" (default), "WGS", "WES", "OMNI".}

\item{datasetId}{(character) options: "gtex_v8" (default), "gtex_v7".}

\item{recordPerChunk}{(integer) number of records fetched per request (default: 200).}

\item{pathologyNotesCategories}{Default: pathologyNotes info is ignored.}
}
\value{
returen sample information
}
\description{

}
\examples{
sampleInfo <- xQTLquery_sampleByTissue(tissueSiteDetail="Liver", datasetId="gtex_v8",
                                       pathologyNotesCategories=TRUE  )

sampleInfo <- xQTLquery_sampleByTissue(tissueSiteDetail="All", dataType="RNASEQ",
                                       datasetId="gtex_v8",pathologyNotesCategories=TRUE )

sampleInfo <- xQTLquery_sampleByTissue("Brain - Amygdala", "RNASEQ","gtex_v8", 200 )
}
