% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{xQTLquery_varId}
\alias{xQTLquery_varId}
\title{Query variant in GTEx with variant ID or dbSNP ID}
\usage{
xQTLquery_varId(variantName = "", variantType = "auto")
}
\arguments{
\item{variantName}{(character) name of variant, dbsnp ID and variant id is supported, eg. "rs138420351" and "chr17_7796745_C_T_b38".}

\item{variantType}{(character) options: "auto", "snpId" or "variantId". Default: "auto".}
}
\value{
A data.table object.
}
\description{
Query variant in GTEx with variant ID or dbSNP ID
}
\examples{
xQTLquery_varId("rs12596338")
xQTLquery_varId("chr11_66561248_T_C_b38")
}
