% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_eqtl}
\alias{xQTLvisual_eqtl}
\title{eQTL significance visualization for a gene}
\usage{
xQTLvisual_eqtl(gene, geneType = "geneSymbol", datasetId = "gtex_v8")
}
\arguments{
\item{gene}{A gene symbol or a gencode id (versioned).}

\item{geneType}{A character string. "geneSymbol"(default) or "gencodeId".}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A plot
}
\description{
eQTL significance visualization for a gene
}
\examples{
\donttest{
  xQTLvisual_eqtl("KIF15")
  xQTLvisual_eqtl("MLH1")
}
}
