% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_geneCorr}
\alias{xQTLvisual_geneCorr}
\title{The correlation plot of two genes’ expression}
\usage{
xQTLvisual_geneCorr(
  gene2 = "",
  geneType = "geneSymbol",
  tissueSiteDetail = "",
  groupBy = "sex",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{gene2}{Gene symbol or gencode ID of two genes. Default: gene symbol.}

\item{geneType}{A character string. "geneSymbol"(default), "gencodeId" or "geneCategory".}

\item{tissueSiteDetail}{Tissue must be accessed by "tissueSiteDetailGTExv7" or "tissueSiteDetailGTExv7".}

\item{groupBy}{Default:sex, can be choosen from pathologyNotesCategories, like: pathologyNotesCategories.mastopathy, pathologyNotesCategories.mastopathy.metaplasia.}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A plot
}
\description{
The correlation plot of two genes’ expression
}
\examples{
\donttest{
 gene2 = c("AMOT", "HOXA4")
 xQTLvisual_geneCorr( gene2, tissueSiteDetail="Liver" )
 xQTLvisual_geneCorr( gene2, groupBy="pathologyNotesCategories.congestion", tissueSiteDetail="Lung" )
}
}
