% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_PNPlot}
\alias{xQTLvisual_PNPlot}
\title{P-N plot}
\usage{
xQTLvisual_PNPlot(summaryDT)
}
\arguments{
\item{summaryDT}{A data.frame with three cols: pval,  beta, se.}
}
\value{
A ggplot2 object
}
\description{
P-N plot
}
\examples{
\donttest{
url1 <- "https://raw.githubusercontent.com/dingruofan/exampleData/master/eqtl/MMP7_qtlDF.txt"
qtl <- fread(url1, sep="\t")
xQTLvisual_PNPlot(qtl[,.(pValue, beta, se)])
}
}
