% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_qqPlot}
\alias{xQTLvisual_qqPlot}
\title{Quantile-quantile plot with p-values from GWAS summary statistics data}
\usage{
xQTLvisual_qqPlot(
  summaryDT,
  legend_p = FALSE,
  binCutLogP = 3,
  binNumber = 1000,
  axis_text_size = 1.3,
  axis_title_size = 1.3,
  title_size = 1.4,
  title_text = "",
  point_color = "#5A90BE"
)
}
\arguments{
\item{summaryDT}{A data.frame of one col required: pval.}

\item{legend_p}{TRUE or FALSE, or legend position, including: top, bottom, left and right.}

\item{binCutLogP}{SNPs whose logP great than this will be binned, other than not binned.}

\item{binNumber}{Number of bins.}

\item{axis_text_size}{(numberic) text size of the axis labels}

\item{axis_title_size}{(numberic) text size of the axis title}

\item{title_size}{(numberic) text size of the title of the plot}

\item{title_text}{(character) title of the plot}

\item{point_color}{(character) Color of the point, default: #5A90BE}
}
\value{
ggplot2 object
}
\description{
Quantile-quantile plot with p-values from GWAS summary statistics data
}
\examples{
\donttest{
url1 <- "http://bioinfo.szbl.ac.cn/xQTL_biolinks/xqtl_data/gwas/gwasSub.txt.gz"
snpInfo <- data.table::fread(url1, sep="\t")
xQTLvisual_qqPlot(snpInfo[,.(pValue)],binCutLogP=5, binNumber=10000)
}
}
