% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{apiAdmin_ping}
\alias{apiAdmin_ping}
\title{Heartbeat to check server connectivity.}
\usage{
apiAdmin_ping()
}
\value{
A numeric value. Response code 200 indicates that the request has succeeded.
}
\description{
test API server
}
\examples{
\donttest{
  apiAdmin_ping()
  ifelse( apiAdmin_ping() ==200, "accessed", "failed")
 }
}
