% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{apiAdmin_ping}
\alias{apiAdmin_ping}
\title{Heartbeat to check GTEx API server connectivity.}
\usage{
apiAdmin_ping(fetchMethod = "")
}
\arguments{
\item{fetchMethod}{fetchMethod.}
}
\value{
A character of fetchContent method.
}
\description{
test GTEx API server and return download method.
}
\examples{
\donttest{
  apiAdmin_ping()
 }
}
