% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{apiAdmin_ping}
\alias{apiAdmin_ping}
\title{Heartbeat to check server connectivity.}
\usage{
apiAdmin_ping()
}
\value{
A boolean value.
}
\description{
test API server
}
\examples{
\donttest{
  apiAdmin_ping()
  apiStatus <- ifelse( apiAdmin_ping() ==200, "GTEx API can be accessed", "Please check your network!")
  print(apiStatus)
 }
}
