% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R, R/internal_dealWithGencode.R
\name{apiAdmin_ping}
\alias{apiAdmin_ping}
\title{Heartbeat to check server connectivity.}
\usage{
apiAdmin_ping()

apiAdmin_ping()
}
\value{
A numeric value. Response code 200 indicates that the request has succeeded.

A character of fetchContent method.

A character of fetchContent method.
}
\description{
test API server and return download method.

test GTEx API server and return download method.
}
\examples{
\donttest{
  apiAdmin_ping()
 }
\donttest{
  apiAdmin_ping()
 }
}
