% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{dbsnpQueryRange}
\alias{dbsnpQueryRange}
\title{retrieve snps from dbSNP using genome range.}
\usage{
dbsnpQueryRange(
  chrom = "",
  startPos = -1,
  endPos = -1,
  genomeBuild = "GRCh38/hg38",
  track = "snp151Common"
)
}
\arguments{
\item{chrom}{A character string. Chromesome of human, including chr1-chr23, chrX, chrY.}

\item{startPos}{A positive integer.}

\item{endPos}{A positive integer.}

\item{genomeBuild}{"GRCh38/hg38" or "GRCh38/hg19". Default: "GRCh38/hg38".}

\item{track}{"snp151Common", "snp150Common" or "snp147Common". Default: "snp151Common".}
}
\value{
A data.table.
}
\description{
retrieve snps from dbSNP using genome range.
}
\examples{
\donttest{
 snpInfo <- dbsnpQueryRange(chrom="chr1", startPos=1, endPos=10000000, genomeBuild="GRCh38/hg38", track="snp151Common" )
}
}
