% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{EBIquery_allTerm}
\alias{EBIquery_allTerm}
\title{EBIquery_allTerm}
\usage{
EBIquery_allTerm(term = "genes", termSize = 5000)
}
\arguments{
\item{term}{"associations", "molecular_phenotypes", "studies", "tissues", "qtl_groups", "genes" or "chromosomes".}
}
\value{
a data.table
}
\description{
EBIquery_allTerm
}
\examples{
\donttest{

 a <- EBIquery_allTerm("associations",termSize=0)
 a <- EBIquery_allTerm("molecular_phenotypes")
 a <- EBIquery_allTerm("studies")
 a <- EBIquery_allTerm("tissues")
 a <- EBIquery_allTerm("qtl_groups")
 a <- EBIquery_allTerm("genes")
 a <- EBIquery_allTerm("chromosomes")
}
}
