% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{EBIquery_allTerm}
\alias{EBIquery_allTerm}
\title{EBIquery_allTerm}
\usage{
EBIquery_allTerm(term = "genes", termSize = 5000)
}
\arguments{
\item{term}{"associations", "molecular_phenotypes", "studies", "tissues", "qtl_groups", "genes" or "chromosomes".}
}
\value{
a data.table
}
\description{
EBIquery_allTerm
}
\examples{
\donttest{

 associations <- rbindlist(EBIquery_allTerm("associations",termSize=0))
 molecular_phenotypes <- EBIquery_allTerm("molecular_phenotypes")
 studies <- EBIquery_allTerm("studies")
 tissues <- EBIquery_allTerm("tissues")
 qtl_groups <- EBIquery_allTerm("qtl_groups")
 genes <- EBIquery_allTerm("genes")
 chromosomes <- EBIquery_allTerm("chromosomes")

 merge(qtl_groups, tissueSiteDetailGTExv8, by.x="qtl_group", by.y="tissueSiteDetail")
}
}
