% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_colocResultAll}
\alias{example_Coloc_colocResultAll}
\title{data for vignette}
\format{
A Data.table
\describe{
\item{nsnps}{A character vector}
\item{PP.H0.abf}{A numeric vector}
\item{PP.H1.abf}{A numeric vector}
\item{PP.H2.abf}{A numeric vector}
\item{PP.H3.abf}{A numeric vector}
\item{PP.H4.abf}{A numeric vector}
\item{traitGene}{A character vector}
}
}
\source{
xQTLanalyze_coloc
}
\description{
head 6 rows of traitsAll of xQTLanalyze_coloc
}
\keyword{internal}
