% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_hyprcolocResultAll}
\alias{example_Coloc_hyprcolocResultAll}
\title{data for vignette}
\format{
A Data.table
\describe{
\item{traitGene}{A character vector}
\item{posterior_prob}{A numeric vector}
\item{regional_prob}{A numeric vector}
\item{candidate_snp}{A character vector}
\item{posterior_explained_by_snp}{A numeric vector}
}
}
\source{
xQTLanalyze_coloc
}
\description{
head 5 rows of hyprcoloc results from xQTLanalyze_coloc
}
\keyword{internal}
