% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_sentinelSNP}
\alias{example_Coloc_sentinelSNP}
\title{data for vignette}
\format{
A Data.table
\describe{
\item{rsid}{A character vector}
\item{chr}{A character vector}
\item{position}{A integer vector}
\item{pValue}{A numeric vector}
\item{maf}{A numeric vector}
}
}
\source{
xQTLanalyze_getSentinelSnp
}
\description{
head 6 rows of sentinelSnpDF of xQTLanalyze_getSentinelSnp
}
\keyword{internal}
