% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{example_Coloc_traitsAll}
\alias{example_Coloc_traitsAll}
\title{data for vignette}
\format{
A Data.table
\describe{
\item{chromosome}{A character vector}
\item{geneStart}{A integer vector}
\item{geneEnd}{A integer vector}
\item{geneStrand}{A character vector}
\item{geneSymbol}{A character vector}
\item{gencodeId}{A character vector}
\item{rsid}{A character vector}
\item{position}{A integer vector}
\item{pValue}{A numeric vector}
\item{maf}{A numeric vector}
}
}
\source{
xQTLanalyze_getTraits
}
\description{
head 6 rows of traitsAll of xQTLanalyze_getTraits
}
\keyword{internal}
