% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{extractGeneInfo}
\alias{extractGeneInfo}
\title{extract gene infor of specified genome from gencodeGeneInfoAllGranges}
\usage{
extractGeneInfo(gencodeGeneInfoAllGranges, genomeVersion = "v26")
}
\arguments{
\item{gencodeGeneInfoAllGranges}{from internal data}

\item{genomeVersion}{"v26" (default) or "v19"}
}
\value{
A data.table object.
}
\description{
extract gene infor of specified genome from gencodeGeneInfoAllGranges
}
\examples{
gencodeGeneInfo <- extractGeneInfo(gencodeGeneInfoAllGranges)
}
