% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fetchContent}
\alias{fetchContent}
\title{Fetch data using url by three methods}
\usage{
fetchContent(
  url1,
  method = "curl",
  downloadMethod = "auto",
  isJson = TRUE,
  retryTimes = 4
)
}
\arguments{
\item{url1}{A url string.}

\item{method}{Can be chosen from "download", "curl", "fromJSON".}

\item{downloadMethod}{The same methods from utils::download.file function.}

\item{isJson}{Fetched content is a json file or not. Defaulst: TRUE.}

\item{retryTimes}{retry times. Default:4}
}
\value{
A json object.
}
\description{
Fetch data using url by three methods
}
\keyword{internal}
