% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{fetchContent}
\alias{fetchContent}
\title{Fetch data using url by three methods}
\usage{
fetchContent(url1, method = "curl", downloadMethod = "auto", isJson = TRUE)
}
\arguments{
\item{url1}{A url string.}

\item{method}{Can be chosen from "download", "curl", "fromJSON", or "GET".}

\item{downloadMethod}{The same methods from utils::download.file function.}

\item{isJson}{Fetched content is a json file or not. Defaulst: TRUE.}
}
\value{
A json object.
}
\description{
Fetch data using url by three methods
}
\examples{
\donttest{
 url1 <- "https://gtexportal.org/rest/v1/admin/ping"
 fetchContent(url1, method="download")
 url1 <- "https://gtexportal.org/rest/v1/association/dyneqtl?gencodeId=ENSG00000065613.13&variantId=chr11_66561248_T_C_b38&tissueSiteDetailId=Liver&datasetId=gtex_v8"

 a <- fetchContent(url1, method="download", isJson=TRUE)

 url1 <- paste0("https://ldlink.nci.nih.gov/LDlinkRest/ldproxy?",
                "var=rs3&pop=MXL&r2_d=r2&window=500000&genome_build=grch38&token=9246d2db7917")
 fetchContent(url1, method="download", isJson=FALSE)
}
}
