% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fetchContent}
\alias{fetchContent}
\title{Fetch data using url by three methods}
\usage{
fetchContent(url1, method = "curl", downloadMethod = "auto", isJson = TRUE)
}
\arguments{
\item{url1}{A url string.}

\item{method}{Can be chosen from "download", "curl", "fromJSON", or "GET".}

\item{downloadMethod}{The same methods from utils::download.file function.}

\item{isJson}{Fetched content is a json file or not. Defaulst: TRUE.}
}
\value{
A json object.
}
\description{
Fetch data using url by three methods
}
\examples{
\donttest{
 url1 <- "https://gtexportal.org/rest/v1/admin/ping"
 fetchContent(url1, method="fromJSON")
}
}
