% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{gencodeGeneInfoAllGranges}
\alias{gencodeGeneInfoAllGranges}
\title{Gene annotations (chr1-chr22).}
\format{
A GRanges object
\describe{
\item{seqnames}{character string, chromosome}
\item{ranges}{Iranges, gene location of v26}
\item{strand}{character string, strand}
\item{rangesV19}{Iranges, gene location of v19}
\item{gencodeId}{character, gencode id}
}
}
\source{
\url{https://gtexportal.org/home/}
}
\description{
A dataset containing the gene information.
}
\keyword{internal}
