% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{gencodeGenetype}
\alias{gencodeGenetype}
\title{Gene types in GTEx V8 and V7}
\format{
A list
\describe{
\item{V26}{A character vector}
\item{V19}{A character vector}
}
}
\source{
\url{https://gtexportal.org/home/}
}
\description{
A dataset containing the classification of genes
}
\keyword{internal}
