% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_coloc}
\alias{GTExanalyze_coloc}
\title{Colocalization analysis with deteched trait}
\usage{
GTExanalyze_coloc(
  gwasDF,
  traitGene,
  sentinelSnp,
  tissueSiteDetail = "Brain - Cortex",
  colocRange = 1e+06,
  mafThreshold = 0.01,
  gwasSampleNum = 50000,
  population = "EUR",
  token = "9246d2db7917",
  method = "coloc"
)
}
\arguments{
\item{gwasDF}{1}

\item{tissueSiteDetail}{1}

\item{population}{1}

\item{token}{1}

\item{method}{1}

\item{traitGenes}{1}
}
\value{
coloc resut
}
\description{
Colocalization analysis with deteched trait
}
\examples{
\donttest{
  traitsAllURL <- "https://gitee.com/stronghoney/exampleData/raw/master/gwas/AD/traitsAll.txt"
  traitsAll <- data.table::fread(rawToChar(curl::curl_fetch_memory(traitsAllURL)$content), sep="\t")
  traitGene <- traitsAll[3,]$geneSymbol
  sentinelSnp <- traitsAll[3,]$rsid

  colocResult <- GTExanalyze_coloc(gwasDF, traitGenes)
  colocResult$gwasEqtlInfo
  # eQTL locuszoom:
  GTExvisual_locusZoom( colocResult$gwasEqtlInfo[,c("rsid","chr","position","pValue.eqtl")], population="EUR",genomeVersion="grch38" )
  # GWAS locuszoom:
  GTExvisual_locusZoom( colocResult$gwasEqtlInfo[,c("rsid","chr","position","pValue.gwas")], population="EUR",genomeVersion="grch38" )
  # locuscompare:
  GTExvisual_locusCompare( colocResult$gwasEqtlInfo[,c("rsid","pValue.eqtl")], colocResult$gwasEqtlInfo[,c("rsid","pValue.gwas")] )
}
}
