% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_coloc}
\alias{GTExanalyze_coloc}
\title{coloc with a gwas dataset}
\usage{
GTExanalyze_coloc(
  gwasDF,
  traitGene = "",
  traitGeneType = "geneSymbol",
  tissueSiteDetail = "",
  study_id = "gtex_v8"
)
}
\arguments{
\item{study_id}{}
}
\value{

}
\description{
coloc with a gwas dataset
}
\examples{
\donttest{
 gwasDF_raw <- data.table::fread("D:\\\\R_project\\\\GWAS_Type-2-Diabetes_Wood_2016.txt.gz", sep="\t", header=TRUE)
 # gwasDF <- gwasDF_raw[,.(rsid, pvalue, beta, varbeta)]
 names(gwasDF) <- c("snpId", "pValue", "beta", "varbeta")
 gwasDF <- gwasDF_raw[,.(rsid, pvalue, Freq)]
 names(gwasDF) <- c("snpId", "pValue", "freq")
}
}
