% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_coloc}
\alias{GTExanalyze_coloc}
\title{coloc analysis with deteched trait}
\usage{
GTExanalyze_coloc(
  gwasDF,
  traitGenes,
  tissueSiteDetail = "Brain - Cortex",
  colocRange = 1e+06,
  numSnp = 3000,
  gwasSampleNum = 50000,
  population = "EUR",
  token = "9246d2db7917",
  method = "coloc"
)
}
\arguments{
\item{gwasDF}{1}

\item{traitGenes}{1}

\item{tissueSiteDetail}{1}

\item{population}{1}

\item{token}{1}

\item{method}{1}
}
\value{
coloc resut
}
\description{
coloc analysis with deteched trait
}
\examples{
\donttest{
  traitGenes <- c("CR1")
  colocResult <- GTExanalyze_coloc(gwasDF, traitGenes)
  colocResult$gwasEqtlInfo
  # eQTL locuszoom:
  GTExvisual_locusZoom( colocResult$gwasEqtlInfo[,c("rsid","chr","position","pValue.eqtl")], population="EUR",genome="grch38" )
  # GWAS locuszoom:
  GTExvisual_locusZoom( colocResult$gwasEqtlInfo[,c("rsid","chr","position","pValue.gwas")], population="EUR",genome="grch38" )
  # locuscompare:
  GTExvisual_locusCompare( colocResult$gwasEqtlInfo[,c("rsid","pValue.eqtl")], colocResult$gwasEqtlInfo[,c("rsid","pValue.gwas")] )
}
}
