% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_coloc}
\alias{GTExanalyze_coloc}
\title{Colocalization analysis with deteched trait}
\usage{
GTExanalyze_coloc(
  gwasDF,
  traitGene,
  geneType = "geneSymbol",
  genomeVersion = "grch38",
  tissueSiteDetail = "",
  colocRange = 0,
  mafThreshold = 0.01,
  gwasSampleNum = 50000,
  method = "coloc"
)
}
\arguments{
\item{gwasDF}{1}

\item{tissueSiteDetail}{1}

\item{method}{1}

\item{traitGenes}{1}

\item{population}{1}

\item{token}{1}
}
\value{
coloc resut
}
\description{
Colocalization analysis with deteched trait
}
\examples{
\donttest{

  traitsAll <- merge(traitsAllEqtl[pValue !="N/A",.(gencodeId, rsid=snpId)], traitsAll, by = c("gencodeId", "rsid"))
  traitsAll <- traitsAll[order(pValue)]

  traitsAllURL <- "https://gitee.com/stronghoney/exampleData/raw/master/gwas/AD/traitsAll.txt"
  traitsAll <- data.table::fread(rawToChar(curl::curl_fetch_memory(traitsAllURL)$content), sep="\t")

  colocResultAll <- list()
  for(i in 1:nrow(traitsAll)){
    message(i," - ", date())
    traitGene <- traitsAll[order(pValue)][i,]$geneSymbol
    colocResult <- GTExanalyze_coloc(gwasDF, traitGene, geneType="geneSymbol", genomeVersion="grch37", tissueSiteDetail="Brain - Cortex")
    colocResultAll[[i]] <- colocResult
  }
  range( unlist(lapply(1:length(colocResultAll), function(x){colocResultAll[[x]]$coloc_Out_summary$PP.H4.abf})) )

  traitGene = traitsAll[geneSymbol=="NUDT17"]$geneSymbol
  colocResult <- GTExanalyze_coloc(gwasDF, traitGene, geneType="geneSymbol", genomeVersion="grch37", tissueSiteDetail="Brain - Cortex")

  colocResult$gwasEqtlInfo
  # eQTL locuszoom:
  gwas_locusRoom <- GTExvisual_locusZoom( colocResult$gwasEqtlInfo[,c("rsid","chr","position","pValue.eqtl")],population="EUR",genomeVersion="grch38" )
  # GWAS locuszoom:
  eQTL_locusRoom <- GTExvisual_locusZoom( colocResult$gwasEqtlInfo[,c("rsid","chr","position","pValue.gwas")], population="EUR",genomeVersion="grch38" )
  # locuscompare:
  locuscompareP <- GTExvisual_locusCompare( colocResult$gwasEqtlInfo[,c("rsid","pValue.eqtl")], colocResult$gwasEqtlInfo[,c("rsid","pValue.gwas")] )
}
}
