% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExanalyze_eqtlGWAS}
\alias{GTExanalyze_eqtlGWAS}
\title{significance analysis of GTEX and eQTL}
\usage{
GTExanalyze_eqtlGWAS(
  gwasDF,
  queryTerm = "",
  queryType = "snpId",
  flankUp = 100,
  flankDown = 100,
  coordinate = "chr1:1-300000",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{queryType}{A character string. for variant, "snpId", "variantId". For gene, "geneSymbol" or "gencodeId".}

\item{flankUp}{A whole integer. KB. Default: 100}

\item{flankDown}{A whole integer. KB. Default: 100}
}
\value{

}
\description{
significance analysis of GTEX and eQTL
}
