% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_getSentinelSnp}
\alias{GTExanalyze_getSentinelSnp}
\title{Sentinel SNPs detection in GWAS data.}
\usage{
GTExanalyze_getSentinelSnp(
  gwasDF,
  pValueThreshold = 5e-08,
  centerRange = 10000,
  mafThreshold = 0.01,
  genomeVersion = "grch38",
  grch37To38 = FALSE
)
}
\arguments{
\item{gwasDF}{GWAS data.frame}
}
\value{
A data.table object.
}
\description{
Return sentinel snps whose pValue < 5e-8(default) and SNP-to-SNP distance > 1e6 bp.
}
\examples{
\donttest{
   gwasFile <- tempfile(pattern = "file")
   gwasURL <- "https://raw.githubusercontent.com/dingruofan/exampleData/master/gwas/AD/GLGC_AD_chr1_6_Sub3.txt"
   utils::download.file(gwasURL, destfile=gwasFile)
   gwasDF <- data.table::fread(gwasFile, sep="\t")
   gwasDF <- gwasDF[, .(rsid, chr, position, P, maf)]
   sentinelSnpDF <- GTExanalyze_getSentinelSnp(gwasDF)

   gwasDF <- fread("D:\\\\R_project\\\\GLGC_CG0052_result.txt.gz", sep="\t")
   gwasDF <- gwasDF[,.(rsid, chr, position, `p-value`, maf)]
   sentinelSnpDF <- GTExanalyze_getSentinelSnp(gwasDF, centerRange=1e4, genomeVersion="grch37", grch37To38=TRUE)
}
}
