% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_traitDetect}
\alias{GTExanalyze_traitDetect}
\title{Sentinel SNPs detection in GWAS data.}
\usage{
GTExanalyze_traitDetect(
  gwasDF,
  pValueThreshold = 5e-08,
  colocRange = 1e+06,
  tissueSiteDetail = "Brain - Cortex",
  maf_threshold = 0.01
)
}
\arguments{
\item{gwasDF}{GWAS data.frame}
}
\value{

}
\description{
Sentinel SNPs detection in GWAS data.
}
\examples{
\donttest{
   gwasFile <- tempfile(pattern = "file")
   gwasURL <- "https://raw.githubusercontent.com/dingruofan/exampleData/master/gwas/AD/GLGC_AD_chr1_6_Sub3.txt"
   utils::download.file(gwasURL, destfile=gwasFile)
   gwasDF <- data.table::fread(gwasFile, sep="\t")
   gwasDF <- gwasDF[, .(rsid, chr, position, P, maf)]
   traits <- GTExanalyze_traitDetect(gwasDF)
}
}
