% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{GTExanalyze_traitEqtlSig}
\alias{GTExanalyze_traitEqtlSig}
\title{Sentinel SNPs detection for a pair of gene and SNP in GTEx .}
\usage{
GTExanalyze_traitEqtlSig(traitGenes, sentinelSnps, tissueSiteDetail = "")
}
\arguments{
\item{traitGenes}{Gene name or genecode ID (unversioned or hg38)}

\item{sentinelSnps}{Variant id (hg38) or dbSNP rsid}

\item{tissueSiteDetail}{A character string.}
}
\value{
A data.table
}
\description{
Sentinel SNPs detection for a pair of gene and SNP in GTEx .
}
\examples{
\donttest{
  gtexEqtl <- GTExanalyze_traitEqtlSig(traitGenes = traitsAll$gencodeId, sentinelSnps = traitsAll$rsid, tissueSiteDetail="Skin - Sun Exposed (Lower leg)" )
}
}
