% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_assoAll}
\alias{GTExdownload_assoAll}
\title{GTExdownload_assoAll}
\usage{
GTExdownload_assoAll(gene = "", geneType = "geneSymbol", tissueSiteDetail = "")
}
\arguments{
\item{gene}{gene}

\item{geneType}{geneType}

\item{tissueSiteDetail}{tissueSiteDetail}
}
\value{
a data.table
}
\description{
GTExdownload_assoAll
}
\examples{
\donttest{
  GTExdownload_assoAll("tp53", tissueSiteDetail="Lung")
  GTExdownload_assoAll("ATP11B", tissueSiteDetail="Muscle - Skeletal")
}
}
