% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_assoAll}
\alias{GTExdownload_assoAll}
\title{GTExdownload_assoAll}
\usage{
GTExdownload_assoAll(
  gene = "",
  geneType = "geneSymbol",
  tissueSiteDetail = "",
  recordPerChunk = 250,
  study = "GTEx_V8",
  withB37VariantId = TRUE
)
}
\arguments{
\item{gene}{gene}

\item{geneType}{geneType}

\item{tissueSiteDetail}{tissueSiteDetail}

\item{recordPerChunk}{A integer value (1-200). number of records fetched per request (default: 200).}

\item{study}{"GTEx_V8"}

\item{withB37VariantId}{Whether to return the genome location(GTEx v7) of variants. Default: TRUE.}
}
\value{
a data.table
}
\description{
GTExdownload_assoAll
}
\examples{
\donttest{
  geneAsso <- GTExdownload_assoAll("ATP11B", tissueSiteDetail="Muscle - Skeletal", withB37VariantId=FALSE)
}
}
