% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_eqtl}
\alias{GTExdownload_eqtl}
\title{Fetch significant eQTL data.}
\usage{
GTExdownload_eqtl(
  variantName = "",
  gene = "",
  variantType = "snpId",
  geneType = "geneSymbol",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{gene}{A gene symbol or a gencode id (versioned).}

\item{variantType}{A character string. "snpId" or "variantId". Default: "snpId".}

\item{geneType}{A character string in "geneSymbol"(default) and "gencodeId".}

\item{tissueSiteDetail}{A character string. Tissue name.}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A data.table
}
\description{
Fetch significant eQTL associations with a variant or a gene or a variant-gene pair in a tissue or across all tissues.
}
\examples{
\donttest{
 # Search with a variant:
 GTExdownload_eqtl(variantName="rs201327123", variantType="snpId")
 GTExdownload_eqtl(variantName="chr1_14677_G_A_b38", variantType="variantId")
 GTExdownload_eqtl(variantName="11_66328719_T_C_b37", variantType="variantId",
                   datasetId="gtex_v7")
 GTExdownload_eqtl(variantName="11_66328719_T_C_b37", variantType="variantId",
                   datasetId="gtex_v7", tissueSiteDetail="Skin - Sun Exposed (Lower leg)")

 # Search with a gene:
 GTExdownload_eqtl(gene="TP53", geneType="geneSymbol")
 GTExdownload_eqtl(gene="ENSG00000141510.16", geneType="gencodeId", datasetId="gtex_v8")
 GTExdownload_eqtl(gene="ENSG00000141510.11", geneType="gencodeId",
                   datasetId="gtex_v7",tissueSiteDetail="Thyroid" )

 # Search with a variant-gene pair:
 GTExdownload_eqtl(variantName="rs1641513",gene="TP53", datasetId="gtex_v8")
 GTExdownload_eqtl(variantName="rs1641513",gene="TP53", datasetId="gtex_v7")
 GTExdownload_eqtl(variantName="chr1_1667948_A_G_b38", variantType="variantId",
                   gene="SLC35E2B", geneType="geneSymbol", tissueSiteDetail="Kidney - Cortex")
}
}
