% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{GTExdownload_geneMedExp}
\alias{GTExdownload_geneMedExp}
\title{Title}
\usage{
GTExdownload_geneMedExp(
  genes = "",
  geneType = "geneSymbol",
  datasetId = "gtex_v8",
  tissueSiteDetail = "",
  recordPerChunk = 150
)
}
\arguments{
\item{recordPerChunk}{}
}
\value{

}
\description{
Title
}
\examples{
\donttest{
 GTExdownload_geneMedExp(genes="TP53", "geneSymbol")
}
}
