% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery_gene.R
\name{GTExquery_gene}
\alias{GTExquery_gene}
\title{query gene info}
\usage{
GTExquery_gene(
  genes,
  geneType = "symbol",
  gencodeVersion = "v26",
  genomeBuild = "GRCh38/hg38"
)
}
\arguments{
\item{genes}{gene names or gene ids (case ignored). Following gene types are supported: gene symbol (character); gencode/ensemble id(versioned or unversioned)(character); entrez gene ID (integer); Default: gene symbol (like: "tp53","naDK","SDF4").}

\item{geneType}{types of queried genes. Options: "symbol", "gencodeId", "entrezId";}

\item{gencodeVersion}{two version are supported: "v26" and "v19"}

\item{genomeBuild}{"GRCh38/hg38" for gencodeVersion "v26", "GRCh37/hg19" for gencodeVersion "v19"}
}
\value{

}
\description{
users can use this function to search gene of interest
}
\examples{
 # test hg38:
 GTExquery_gene("TP53", "symbol", "v26", "GRCh38/hg38")
 GTExquery_gene(c("tp53","naDK","SDF4"), "symbol", "v26", "GRCh38/hg38")
 GTExquery_gene(c("ENSG00000210195.2","ENSG00000078808","Ensg00000008130.15"), "symbol", "v26", "GRCh38/hg38")
 GTExquery_gene(c(51150,5590,4509), "entrezId", "v26", "GRCh38/hg38")
 # test hg19:
 GTExquery_gene(c(51150,5590,4509), "entrezId", "v19","GRCh37/hg19")
}
