% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{GTExquery_gene}
\alias{GTExquery_gene}
\title{Query gene information}
\usage{
GTExquery_gene(
  genes = "",
  geneType = "symbol",
  gencodeVersion = "v26",
  genomeBuild = "GRCh38/hg38"
)
}
\arguments{
\item{genes}{Following gene types are supported:
\itemize{
\item \strong{Gene symbol}.

A character string or a character vector (case ignored). like: "tp53","naDK","SDF4".
\item \strong{Gencode/ensemble id} (versioned or unversioned).

A character string or a character vector (case ignored). like: "ENSG00000210195.2","ENSG00000078808".
\item \strong{Entrez gene ID}.

A integer string or a integer vectors. like: 51150,5590,4509.
}}

\item{geneType}{A character string. Types of queried genes. Options: "symbol" (default), "gencodeId", "entrezId";}

\item{gencodeVersion}{A character string. Two version are supported: "v26" (default) and "v19"}

\item{genomeBuild}{A character string. "GRCh38/hg38"(default) for gencodeVersion "v26", "GRCh37/hg19" for gencodeVersion "v19"}
}
\value{
A data.table of queried gene information. including following columns:
\itemize{
\item \strong{genes.} Input genes
\item \strong{geneSymbol.} Gene symbol.
\item \strong{gencodeId.} Gencode/ensemble id (versioned).
\item \strong{entrezGeneId.} Entrez gene ID.
\item \strong{geneType.} Gene type.
\item \strong{chromosome.} Note: "chr" is added in gencode v26,
\item \strong{start.}
\item \strong{end.}
\item \strong{strand.}
\item \strong{tss.} Transcriptional start site.
\item \strong{gencodeVersion.} Gencode Version.
\item \strong{genomeBuild.} Genome version.
\item \strong{description.}
}
}
\description{
users can use this function to search gene of interest
}
\examples{
\donttest{
 # test hg38:
 geneInfo <- GTExquery_gene("TP53", "symbol", "v26", "GRCh38/hg38")
 geneInfo <- GTExquery_gene(c("tp53","naDK","SDF4"), "symbol", "v26", "GRCh38/hg38")
 geneInfo <- GTExquery_gene(c("ENSG00000210195.2","ENSG00000078808"), geneType="gencodeId", "v26", "GRCh38/hg38")
 geneInfo <- GTExquery_gene(c(51150,5590,4509), "entrezId", "v26", "GRCh38/hg38")
 # test hg19:
 geneInfo <- GTExquery_gene(c(51150,5590,4509), "entrezId", "v19","GRCh37/hg19")
 }
}
