% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{GTExquery_sample}
\alias{GTExquery_sample}
\title{Fetch information of samples used in analyses from all datasets.}
\usage{
GTExquery_sample(
  tissueSiteDetail = "Liver",
  dataType = "RNASEQ",
  datasetId = "gtex_v8",
  pageSize = 200
)
}
\arguments{
\item{tissueSiteDetail}{For GTEx v8, must be following terms:
\itemize{
\item All
\item Adipose - Subcutaneous
\item Adipose - Visceral (Omentum)
\item Adrenal Gland
\item Artery - Aorta
\item Artery - Coronary
\item Artery - Tibial
\item Bladder
\item Brain - Amygdala
\item Brain - Anterior cingulate cortex (BA24)
\item Brain - Caudate (basal ganglia)
\item Brain - Cerebellar Hemisphere
\item Brain - Cerebellum
\item Brain - Cortex
\item Brain - Frontal Cortex (BA9)
\item Brain - Hippocampus
\item Brain - Hypothalamus
\item Brain - Nucleus accumbens (basal ganglia)
\item Brain - Putamen (basal ganglia)
\item Brain - Spinal cord (cervical c-1)
\item Brain - Substantia nigra
\item Breast - Mammary Tissue
\item Cells - Cultured fibroblasts
\item Cells - EBV-transformed lymphocytes
\item Cervix - Ectocervix
\item Cervix - Endocervix
\item Colon - Sigmoid
\item Colon - Transverse
\item Esophagus - Gastroesophageal Junction
\item Esophagus - Mucosa
\item Esophagus - Muscularis
\item Fallopian Tube
\item Heart - Atrial Appendage
\item Heart - Left Ventricle
\item Kidney - Cortex
\item Kidney - Medulla
\item Liver
\item Lung
\item Minor Salivary Gland
\item Muscle - Skeletal
\item Nerve - Tibial
\item Ovary
\item Pancreas
\item Pituitary
\item Prostate
\item Skin - Not Sun Exposed (Suprapubic)
\item Skin - Sun Exposed (Lower leg)
\item Small Intestine - Terminal Ileum
\item Spleen
\item Stomach
\item Testis
\item Thyroid
\item Uterus
\item Vagina
\item Whole Blood
}
For GTEx v7, must be following terms:
\itemize{
\item All
\item Adipose - Subcutaneous
\item Adipose - Visceral (Omentum)
\item Adrenal Gland
\item Artery - Aorta
\item Artery - Coronary
\item Artery - Tibial
\item Bladder
\item Brain - Amygdala
\item Brain - Anterior cingulate cortex (BA24)
\item Brain - Caudate (basal ganglia)
\item Brain - Cerebellar Hemisphere
\item Brain - Cerebellum
\item Brain - Cortex
\item Brain - Frontal Cortex (BA9)
\item Brain - Hippocampus
\item Brain - Hypothalamus
\item Brain - Nucleus accumbens (basal ganglia)
\item Brain - Putamen (basal ganglia)
\item Brain - Spinal cord (cervical c-1)
\item Brain - Substantia nigra
\item Breast - Mammary Tissue
\item Cells - EBV-transformed lymphocytes
\item Cells - Transformed fibroblasts
\item Cervix - Ectocervix
\item Cervix - Endocervix
\item Colon - Sigmoid
\item Colon - Transverse
\item Esophagus - Gastroesophageal Junction
\item Esophagus - Mucosa
\item Esophagus - Muscularis
\item Fallopian Tube
\item Heart - Atrial Appendage
\item Heart - Left Ventricle
\item Kidney - Cortex
\item Liver
\item Lung
\item Minor Salivary Gland
\item Muscle - Skeletal
\item Nerve - Tibial
\item Ovary
\item Pancreas
\item Pituitary
\item Prostate
\item Skin - Not Sun Exposed (Suprapubic)
\item Skin - Sun Exposed (Lower leg)
\item Small Intestine - Terminal Ileum
\item Spleen
\item Stomach
\item Testis
\item Thyroid
\item Uterus
\item Vagina
\item Whole Blood
}}

\item{dataType}{A character string. Options: "RNASEQ" (default), "WGS", "WES", "OMNI".}

\item{datasetId}{A character string. Options: "gtex_v8" (default), "gtex_v7".}

\item{pageSize}{A integer value (1-2000). number of records fetched per request (default: 200).}
}
\value{

}
\description{
Fetch information of samples used in analyses from all datasets.
}
\examples{
 GTExquery_sample( tissueSiteDetail="Liver", dataType="RNASEQ", datasetId="gtex_v8", pageSize=200 )
 GTExquery_sample( tissueSiteDetail="All", dataType="RNASEQ", datasetId="gtex_v8",pageSize=2000 )
 GTExquery_sample( tissueSiteDetail="Adipose - Visceral (Omentum)", dataType="RNASEQ", datasetId="gtex_v8",pageSize=200 )
}
