% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_eqtlGWAS}
\alias{GTExvisual_eqtlGWAS}
\title{significance analysis of GTEX and eQTL  NO.}
\usage{
GTExvisual_eqtlGWAS(
  gwasDF,
  traitGene = "",
  traitGeneType = "geneSymbol",
  tissueSiteDetail = "",
  study_id = "gtex_v8"
)
}
\arguments{
\item{gwasDF}{A data.frame containing GWAS summary info. At least four columns are expected, including "snpId", "chrom", "pos" and "pValue".}

\item{traitGene}{eqtl trait. If this parameter is not specified, it is the same as the "queryTerm".}

\item{traitGeneType}{eqtl type. If this parameter is not specified, it is the same as the "queryType".}

\item{tissueSiteDetail}{tissue}

\item{study_id}{A character string. Default: "gtex_v8".}
}
\value{
A plot.
}
\description{
significance analysis of GTEX and eQTL  NO.
}
\examples{
\donttest{
 gwasDF_raw <- data.table::fread(
 "D:\\\\R_project\\\\GWAS_Type-2-Diabetes_Wood_2016.txt.gz", sep="\t", header=TRUE)
 gwasDF <- gwasDF_raw[,.(rsid, pvalue, Freq)]
 names(gwasDF) <- c("snpId","pValue","freq")

 traitGene="ABCB9"
 tissueSiteDetail="Whole Blood"
 geneAsso<- GTExdownload_assoAll(traitGene, tissueSiteDetail=tissueSiteDetail)
 gwas_eqtl <- merge(gwasDF, geneAsso, by="snpId", sort=FALSE, suffixes = c(".gwas",".etql"))
 colocResult <- coloc::coloc.abf(dataset1=list(
   pvalues=gwas_eqtl$pValue.gwas, type="cc", s=0.33, N=nrow(gwasDF_raw)),
 dataset2=list(
   pvalues=gwas_eqtl$pValue.etql, type="quant", N=nrow(geneAsso)),
   MAF=gwas_eqtl$maf)

}
}
