% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_eqtlTrait}
\alias{GTExvisual_eqtlTrait}
\title{eql trait plot}
\usage{
GTExvisual_eqtlTrait(
  gene = "",
  geneType = "geneSymbol",
  highlightSnp = "",
  tissueSiteDetail = "",
  study = "gtex_v8",
  population = "ALL"
)
}
\arguments{
\item{gene}{gene name/gencode ID}

\item{geneType}{"geneSymbol" or "gencodeId"}

\item{highlightSnp}{highlighted SNP}

\item{tissueSiteDetail}{gene name}

\item{study}{"gtex_v8"}

\item{population}{(string) One of the 5 popuations from 1000 Genomes: 'AFR', 'AMR', 'EAS', 'EUR', and 'SAS'. Default: 'ALL'.}
}
\value{
A data.table
}
\description{
eql trait plot
}
\examples{
\donttest{
 GTExvisual_eqtlTrait("tp53", tissueSiteDetail = "Lung", population="ALL")
 GTExvisual_eqtlTrait("ENSG00000103642", "gencodeId",
   tissueSiteDetail = "Liver", study="gtex_v8", population=c("EAS","SAS","AMR"))
}
}
