% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_eqtlTrait}
\alias{GTExvisual_eqtlTrait}
\title{eql trait plot}
\usage{
GTExvisual_eqtlTrait(
  gene = "",
  geneType = "geneSymbol",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{gene}{gene name/gencode ID}

\item{geneType}{"geneSymbol" or "gencodeId"}

\item{tissueSiteDetail}{gene name}

\item{datasetId}{"gtex_v7" or "gtex_v8"}
}
\value{
A data.table
}
\description{
eql trait plot
}
\examples{
\donttest{
 GTExvisual_eqtlTrait("ENSG00000112137.12", "gencodeId",tissueSiteDetail = "Adipose - Subcutaneous", datasetId="gtex_v7")
 GTExvisual_eqtlTrait("tp53",tissueSiteDetail = "Adipose - Subcutaneous", datasetId="gtex_v7")
}
}
