% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_eqtlTrait}
\alias{GTExvisual_eqtlTrait}
\title{}
\usage{
GTExvisual_eqtlTrait(
  gene = "",
  geneType = "geneSymbol",
  coordinate = "chr1:1-3",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\value{

}
\description{

}
