% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{GTExvisual_locusZoom}
\alias{GTExvisual_locusZoom}
\title{LocusZoom plot}
\usage{
GTExvisual_locusZoom(
  DF,
  highlightSnp = "",
  population = "EUR",
  range = "",
  token = "9246d2db7917"
)
}
\arguments{
\item{highlightSnp}{Default is the snp that with lowest p-value.}

\item{token}{}

\item{dataFrame}{A data.frame or a data.table object. Two columns are required: "snps", a character list, using an rsID or chromosome coordinate (e.g. "chr7:24966446"); P-value.}
}
\value{

}
\description{
LocusZoom plot
}
\examples{
\donttest{
 eqtlAsso <- GTExdownload_assoAll("MLH1", tissueSiteDetail = "Liver")
 DF <- eqtlAsso[,.(snpId, pValue)]
}
}
