% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dealWithGencode.R
\name{gtfSubsGeneInfo}
\alias{gtfSubsGeneInfo}
\title{download gtf files and extract gene info from attribute column:}
\usage{
gtfSubsGeneInfo(gencodeVersion = "v26")
}
\arguments{
\item{gencodeVersion}{"v26" or "v19"}
}
\value{
create .rds file
}
\description{
download gtf files and extract gene info from attribute column:
}
