% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{determine is a whole number:}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A number}

\item{tol}{Don't change}
}
\value{
TRUE or FALSE
}
\description{
determine is a whole number:
}
