% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissueSiteDetailGTExv7}
\alias{tissueSiteDetailGTExv7}
\title{Tissue name and tissue id mapping of GTEx V7.}
\format{
A data frame with 53 rows and 2 variables:
\describe{
\item{tissueSiteDetail}{character string, tissue name}
\item{tissueSiteDetailId}{character string, tissue id removding special character}
...
}
}
\source{
\url{https://gtexportal.org/home/}
}
\usage{
tissueSiteDetailGTExv7
}
\description{
A dataset containing the 54 tissues' name and corresponding ID.
}
\keyword{datasets}
