% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTLbiolinks.R
\docType{data}
\name{tissueSiteDetailGTExv8}
\alias{tissueSiteDetailGTExv8}
\title{Tissue name and tissue id mapping of GTEx V8.}
\format{
A data frame with 54 rows and 2 variables
\describe{
\item{tissueSiteDetail}{character string, tissue name}
\item{tissueSiteDetailId}{character string, tissue id removding special character}
\item{tissueSite}{character string, whole tissue name}
}
Compared with GTEx v7
\tabular{rrrrr}{
\strong{Tissue name} \tab \strong{GTEx V8} \tab \strong{GTEx V7} \cr
Adipose - Subcutaneous \tab √ \tab √\cr
Adipose - Visceral (Omentum) \tab √ \tab √\cr
Adrenal Gland \tab √ \tab √\cr
Artery - Aorta \tab √ \tab √\cr
Artery - Coronary \tab √ \tab √\cr
Artery - Tibial \tab √ \tab √\cr
Bladder \tab √ \tab √\cr
Brain - Amygdala \tab √ \tab √\cr
Brain - Anterior cingulate cortex (BA24) \tab √ \tab √\cr
Brain - Caudate (basal ganglia) \tab √ \tab √\cr
Brain - Cerebellar Hemisphere \tab √ \tab √\cr
Brain - Cerebellum \tab √ \tab √\cr
Brain - Cortex \tab √ \tab √\cr
Brain - Frontal Cortex (BA9) \tab √ \tab √\cr
Brain - Hippocampus \tab √ \tab √\cr
Brain - Hypothalamus \tab √ \tab √\cr
Brain - Nucleus accumbens (basal ganglia) \tab √ \tab √\cr
Brain - Putamen (basal ganglia) \tab √ \tab √\cr
Brain - Spinal cord (cervical c-1) \tab √ \tab √\cr
Brain - Substantia nigra \tab √ \tab √\cr
Breast - Mammary Tissue \tab √ \tab √\cr
Cells - Cultured fibroblasts \tab √ \tab x\cr
Cells - EBV-transformed lymphocytes \tab √ \tab √\cr
Cells - Transformed fibroblasts \tab x \tab √\cr
Cervix - Ectocervix \tab √ \tab √\cr
Cervix - Endocervix \tab √ \tab √\cr
Colon - Sigmoid \tab √ \tab √\cr
Colon - Transverse \tab √ \tab √\cr
Esophagus - Gastroesophageal Junction \tab √ \tab √\cr
Esophagus - Mucosa \tab √ \tab √\cr
Esophagus - Muscularis \tab √ \tab √\cr
Fallopian Tube \tab √ \tab √\cr
Heart - Atrial Appendage \tab √ \tab √\cr
Heart - Left Ventricle \tab √ \tab √\cr
Kidney - Cortex \tab √ \tab √\cr
Kidney - Medulla \tab √ \tab x\cr
Liver \tab √ \tab √\cr
Lung \tab √ \tab √\cr
Minor Salivary Gland \tab √ \tab √\cr
Muscle - Skeletal \tab √ \tab √\cr
Nerve - Tibial \tab √ \tab √\cr
Ovary \tab √ \tab √\cr
Pancreas \tab √ \tab √\cr
Pituitary \tab √ \tab √\cr
Prostate \tab √ \tab √\cr
Skin - Not Sun Exposed (Suprapubic) \tab √ \tab √\cr
Skin - Sun Exposed (Lower leg) \tab √ \tab √\cr
Small Intestine - Terminal Ileum \tab √ \tab √\cr
Spleen \tab √ \tab √\cr
Stomach \tab √ \tab √\cr
Testis \tab √ \tab √\cr
Thyroid \tab √ \tab √\cr
Uterus \tab √ \tab √\cr
Vagina \tab √ \tab √\cr
Whole Blood \tab √ \tab √\cr
}
}
\source{
\url{https://gtexportal.org/home/}
}
\description{
A dataset containing the 54 tissues' name and corresponding ID of GTEx V8.
}
\keyword{internal}
