% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_coloc}
\alias{xQTLanalyze_coloc}
\title{xQTLanalyze_coloc}
\usage{
xQTLanalyze_coloc(
  gwasDF,
  traitGene,
  geneType = "auto",
  genomeVersion = "grch38",
  tissueSiteDetail = "",
  mafThreshold = 0.01,
  population = "EUR",
  gwasSampleNum = 50000,
  method = "coloc",
  token = "9246d2db7917"
)
}
\arguments{
\item{gwasDF}{A dataframe of gwas.}

\item{traitGene}{A gene symbol or a gencode id (versioned).}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{genomeVersion}{"grch38" (default) or "grch37". Note: grch37 will be converted to grch38 automatically.}

\item{tissueSiteDetail}{A character string. Tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{mafThreshold}{Cutoff of maf to remove rare variants.}

\item{population}{Supported population is consistent with the LDlink, which can be listed using function "LDlinkR::list_pop()"}

\item{gwasSampleNum}{Sample number of GWAS dataset. Default:50000.}

\item{method}{Now only one "coloc". Package \code{coloc} is required.}

\item{token}{LDlink provided user token, default = NULL, register for token at https://ldlink.nci.nih.gov/?tab=apiaccess}
}
\value{
coloc resut
}
\description{
conduct colocalization analysis with detected gene.
}
\examples{
\donttest{
  # please see see vignette.
}
}
