% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{xQTLanalyze_coloc}
\alias{xQTLanalyze_coloc}
\title{xQTLanalyze_coloc}
\usage{
xQTLanalyze_coloc(
  gwasDF,
  traitGene,
  geneType = "auto",
  genomeVersion = "grch38",
  tissueSiteDetail = "",
  mafThreshold = 0.01,
  gwasSampleNum = 50000,
  method = "coloc"
)
}
\arguments{
\item{gwasDF}{A dataframe of gwas.}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{A character string. Tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{method}{Now only one "coloc". Package \code{coloc} is required.}

\item{traitGenes}{A gene symbol or a gencode id (versioned).}

\item{population}{Supported population is consistent with the LDlink, which can be listed using function "LDlinkR::list_pop()"}

\item{token}{LDlink provided user token, default = NULL, register for token at https://ldlink.nci.nih.gov/?tab=apiaccess}
}
\value{
coloc resut
}
\description{
conduct colocalization analysis with detected gene.
}
\examples{
\donttest{

  traitsAll <- merge(traitsAllEqtl[pValue !="N/A",.(gencodeId, rsid=snpId)], traitsAll, by = c("gencodeId", "rsid"))
  traitsAll <- traitsAll[order(pValue)]

  traitsAllURL <- "https://gitee.com/stronghoney/exampleData/raw/master/gwas/AD/traitsAll.txt"
  traitsAll <- data.table::fread(rawToChar(curl::curl_fetch_memory(traitsAllURL)$content), sep="\t")

  colocResultAll <- list()
  for(i in 1:nrow(traitsAll)){
    message(i," - ", date())
    traitGene <- traitsAll[order(pValue)][i,]$geneSymbol
    colocResult <- xQTLanalyze_coloc(gwasDF, traitGene, geneType="geneSymbol", genomeVersion="grch37", tissueSiteDetail="Brain - Cortex")
    colocResultAll[[i]] <- colocResult
  }
  range( unlist(lapply(1:length(colocResultAll), function(x){colocResultAll[[x]]$coloc_Out_summary$PP.H4.abf})) )

  traitGene = traitsAll[geneSymbol=="NUDT17"]$geneSymbol
  colocResult <- xQTLanalyze_coloc(gwasDF, traitGene, geneType="geneSymbol", genomeVersion="grch37", tissueSiteDetail="Brain - Cortex")

  colocResult$gwasEqtlInfo
  # eQTL locuszoom:
  gwas_locusRoom <- xQTLvisual_locusZoom( colocResult$gwasEqtlInfo[,c("rsid","chr","position","pValue.eqtl")],population="EUR",genomeVersion="grch38" )
  # GWAS locuszoom:
  eQTL_locusRoom <- xQTLvisual_locusZoom( colocResult$gwasEqtlInfo[,c("rsid","chr","position","pValue.gwas")], population="EUR",genomeVersion="grch38" )
  # locuscompare:
  locuscompareP <- xQTLvisual_locusCompare( colocResult$gwasEqtlInfo[,c("rsid","pValue.eqtl")], colocResult$gwasEqtlInfo[,c("rsid","pValue.gwas")] )
}
}
