% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{xQTLanalyze_getTraits}
\alias{xQTLanalyze_getTraits}
\title{xQTLanalyze_getTraits}
\usage{
xQTLanalyze_getTraits(
  sentinelSnpDF,
  detectRange = 1e+06,
  tissueSiteDetail = "",
  genomeVersion = "grch38",
  grch37To38 = FALSE
)
}
\arguments{
\item{sentinelSnpDF}{A data.table. Better be the results from the function "xQTLanalyze_getSentinelSnp", five columns are required, including "rsid", "chr", "position", "pValue", and "maf".}

\item{detectRange}{A integer value. Trait genes that harbor sentinel SNPs located in the 1kb range upstream and downstream of gene. Default: 1e6 bp}

\item{tissueSiteDetail}{All tissues' name can be listed with \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{genomeVersion}{"grch38" or "grch37". Default: "grch38"}

\item{grch37To38}{TRUE or FALSE, we recommend converting grch37 to grch38, or using a input file of grch38 directly. Package \code{rtracklayer} is required.}
}
\value{
A data.table object
}
\description{
identify trait genes with sentinel SNPs:
}
\examples{
\donttest{
  URL1<-"https://gitee.com/stronghoney/exampleData/raw/master/gwas/GLGC_CG0052/sentinelSnpDF.txt"
  sentinelSnpDF <- data.table::fread(URL1)
  traitsAll <- xQTLanalyze_getTraits(sentinelSnpDF,detectRange=1e4,"Brain - Cerebellum",
                                     genomeVersion="grch37", grch37To38=TRUE)
}
}
