% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{xQTLanalyze_getTraits}
\alias{xQTLanalyze_getTraits}
\title{Detect the genes around the sentinel SNPs}
\usage{
xQTLanalyze_getTraits(
  sentinelSnpDF,
  detectRange = 10000,
  genomeVersion = "grch38",
  grch37To38 = FALSE
)
}
\arguments{
\item{sentinelSnpDF}{A data.table. Better be the results from the function "xQTLanalyze_getSentinelSnp", five columns are required, including "rsid", "chr", "position", "pValue", and "maf".}

\item{detectRange}{A integer value. Trait genes that harbor sentinel SNPs located in the 1kb range upstream and downstream of gene. Default: 1e6 bp}

\item{genomeVersion}{"grch38" or "grch19". Default: "grch38"}
}
\value{
A data.table object
}
\description{
Detect the genes around the sentinel SNPs
}
\examples{
\donttest{
  sentinelSnpsURL <- "https://gitee.com/stronghoney/exampleData/raw/master/gwas/GLGC_CG0052/sentinelSnpDF.txt"
  sentinelSnpDF <- data.table::fread(rawToChar(curl::curl_fetch_memory(sentinelSnpsURL)$content), sep="\t")
  traitsAll <- xQTLanalyze_getTraits(sentinelSnpDF, detectRange=1e4, genomeVersion="grch37", grch37To38=TRUE)
}
}
