% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExanalyze.R
\name{xQTLanalyze_TSExp}
\alias{xQTLanalyze_TSExp}
\title{xQTLanalyze_TSExp}
\usage{
xQTLanalyze_TSExp(
  genes,
  geneType = "auto",
  method = "SPM",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{geneType}{A character string. "auto"(default), "geneSymbol", "gencodeId" or "geneCategory".}

\item{method}{"SPM" or "entropy"}

\item{datasetId}{"gtex_v8" or "gtex_v7".}

\item{A}{charater vector or a string of gene symbol, gencode id (versioned), or a charater string of gene type.}
}
\value{
A data.table
}
\description{
perform tissue-specific expression analysis.
}
\examples{
\donttest{
 protein_coding <- xQTLquery_gene(genes="protein coding", geneType="geneCategory", "v26" )
 TSgene <- xQTLanalyze_TSExp( unique(protein_coding$gencodeId)[1:100])
 genes <- extractGeneInfo(gencodeGeneInfoAllGranges)$gencodeId[400:410]
 TSgene <- xQTLanalyze_TSExp(genes, datasetId="gtex_v7")
 xQTLvisual_geneExpTissues( TSgene[order(-DPM)][1,]$geneSymbol )
}
}
