% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_egene}
\alias{xQTLdownload_egene}
\title{Download eGenes (eQTL Genes) for a specified gene or a tissue}
\usage{
xQTLdownload_egene(
  tissueSiteDetail = "",
  gene = "",
  geneType = "auto",
  recordPerChunk = 2000
)
}
\arguments{
\item{tissueSiteDetail}{(character) details of tissues in GTEx can be listed using \code{tissueSiteDetailGTExv8} or \code{tissueSiteDetailGTExv7}}

\item{gene}{(character) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{recordPerChunk}{(integer) number of records fetched per request (default: 2000).}
}
\value{
A data.table object.
}
\description{
eGenes are genes that have at least one significant cis-eQTL acting upon them. Results can be filtered by tissue.
}
\examples{
\donttest{
eGeneInfo <- xQTLdownload_egene(tissueSiteDetail="Prostate", gene="CICP3")
eGeneInfo <- xQTLdownload_egene(tissueSiteDetail="Prostate")
}
}
