% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_egene}
\alias{xQTLdownload_egene}
\title{Download details of eGenes (eQTL Genes) for a specified gene or a tissue.}
\usage{
xQTLdownload_egene(
  gene = "",
  geneType = "auto",
  datasetId = "gtex_v8",
  tissueSiteDetail = "",
  recordPerChunk = 200
)
}
\arguments{
\item{gene}{(character) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{(character) options: "gtex_v8" (default), "gtex_v7".}

\item{tissueSiteDetail}{(character) details of tissues in GTEx can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{recordPerChunk}{(integer) number of records fetched per request (default: 200).}
}
\value{
A data.table object.
}
\description{
eGenes are genes that have at least one significant cis-eQTL acting upon them. Results can be filtered by tissue.
}
\examples{
# eGeneInfoAlltissue <- xQTLdownload_egene()
eGeneInfo <- xQTLdownload_egene("TP53")
eGeneInfo <- xQTLdownload_egene(tissueSiteDetail="Lung", recordPerChunk=2000)
eGeneInfo <- xQTLdownload_egene("DDX11", datasetId="gtex_v7", tissueSiteDetail="Artery - Tibial")
}
