% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_egene}
\alias{xQTLdownload_egene}
\title{xQTLdownload_egene}
\usage{
xQTLdownload_egene(
  gene = "",
  geneType = "auto",
  datasetId = "gtex_v8",
  tissueSiteDetail = "",
  recordPerChunk = 200
)
}
\arguments{
\item{gene}{A charater string of gene symbol, gencode id (versioned).}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}

\item{tissueSiteDetail}{A character string. Tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{recordPerChunk}{A integer value (1-200). number of records fetched per request (default: 200).}
}
\value{
a data.table
}
\description{
download eGenes (eQTL Genes).
eGenes are genes that have at least one significant cis-eQTL acting upon them. Results may be filtered by tissue.
}
\examples{
\donttest{
 eGeneInfoAlltissue <- xQTLdownload_egene()
 eGeneInfo <- xQTLdownload_egene("TP53")
 eGeneInfo <- xQTLdownload_egene(tissueSiteDetail="Lung", recordPerChunk=2000)
 eGeneInfo <- xQTLdownload_egene("ENSG00000141510.16")
 eGeneInfo <- xQTLdownload_egene("DDX11", datasetId="gtex_v7", tissueSiteDetail="Artery - Tibial" )
}
}
