% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_eqtlAllAsso}
\alias{xQTLdownload_eqtlAllAsso}
\title{xQTLdownload_eqtlAllAsso}
\usage{
xQTLdownload_eqtlAllAsso(
  gene = "",
  geneType = "auto",
  variantName = "",
  variantType = "auto",
  tissueLabel = "",
  study = "gtex_v8",
  recordPerChunk = 250,
  withB37VariantId = TRUE
)
}
\arguments{
\item{gene}{gene A gene symbol or a gencode id (versioned).}

\item{geneType}{geneType A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{tissueLabel}{A character string, all supported tissues can be listed using "ebi_study_tissues"}

\item{study}{Studies can be listed using "ebi_study_tissues"}

\item{recordPerChunk}{A integer value (1-200). number of records fetched per request (default: 200).}

\item{withB37VariantId}{Whether to return the genome location(GTEx v7) of variants. Default: TRUE.}
}
\value{
a data.table
}
\description{
download all tested variant-gene associations.
source of all eQTL associations is EBI eQTL category.
}
\examples{
geneAsso <- xQTLdownload_eqtlAllAsso("ATP11B", tissueLabel="Muscle - Skeletal",
                                     withB37VariantId=FALSE)
geneAsso1 <- xQTLdownload_eqtlAllAsso(gene= "ATP11B", variantName="rs263022",
                                      withB37VariantId=FALSE)
}
