% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_eqtlAllAsso}
\alias{xQTLdownload_eqtlAllAsso}
\title{xQTLdownload_eqtlAllAsso}
\usage{
xQTLdownload_eqtlAllAsso(
  gene = "",
  geneType = "auto",
  tissueSiteDetail = "",
  recordPerChunk = 250,
  study = "gtex_v8",
  withB37VariantId = TRUE
)
}
\arguments{
\item{gene}{gene A gene symbol or a gencode id (versioned).}

\item{geneType}{geneType A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{tissueSiteDetail A character string. tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{recordPerChunk}{A integer value (1-200). number of records fetched per request (default: 200).}

\item{study}{"gtex_v8" only GTEx v8 is supported.}

\item{withB37VariantId}{Whether to return the genome location(GTEx v7) of variants. Default: TRUE.}
}
\value{
a data.table
}
\description{
download all tested variant-gene associations.
source of all eQTL associations is EBI eQTL category.
}
\examples{
\donttest{
  geneAsso <- xQTLdownload_eqtlAllAsso("ATP11B",
                                       tissueSiteDetail="Muscle - Skeletal",
                                       withB37VariantId=FALSE)
}
}
