% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_eqtlExp}
\alias{xQTLdownload_eqtlExp}
\title{Download normalized expression of gene for a eQTL pair.}
\usage{
xQTLdownload_eqtlExp(
  variantName = "",
  gene = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{variantName}{(character) name of variant, dbsnp ID and variant id is supported, eg. "rs138420351" and "chr17_7796745_C_T_b38".}

\item{gene}{(character) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{variantType}{(character) options: "auto", "snpId" or "variantId". Default: "auto".}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{(character) details of tissues in GTEx can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{datasetId}{(character) options: "gtex_v8" (default), "gtex_v7".}
}
\value{
A data.table object.
}
\description{
Download normalized expression of gene for a eQTL pair.
}
\examples{
# Download exp with variant-gene pair in different tissues:
xQTLdownload_eqtlExp(variantName="rs1641513",gene="TP53", tissueSiteDetail="Liver")
xQTLdownload_eqtlExp(variantName="rs1641513",gene="ATAD3B",
                     tissueSiteDetail="Lung", datasetId="gtex_v8")

# Download exp in gtex v7:
xQTLdownload_eqtlExp(variantName="rs140894808",gene="ATAD3B",
                     tissueSiteDetail="Adipose - Visceral (Omentum)",
                     datasetId="gtex_v7")

# Download expression using variant ID and gencode ID.
xQTLdownload_eqtlExp(variantName="chr1_14677_G_A_b38",gene="ENSG00000228463.9",
                    tissueSiteDetail="Stomach")
}
