% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_eqtlExp}
\alias{xQTLdownload_eqtlExp}
\title{xQTLdownload_eqtlExp}
\usage{
xQTLdownload_eqtlExp(
  variantName = "",
  gene = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{gene}{A gene symbol or a gencode id (versioned).}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{A character string. tissue detail can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A data.table
}
\description{
download normalized expression of gene for a eQTL pair.
}
\examples{
# Download exp with variant-gene pair in different tissues:
xQTLdownload_eqtlExp(variantName="rs1641513",gene="TP53", tissueSiteDetail="Liver")
xQTLdownload_eqtlExp(variantName="rs1641513",gene="ATAD3B",
                     tissueSiteDetail="Lung", datasetId="gtex_v8")

# Download exp in gtex v7:
xQTLdownload_eqtlExp(variantName="rs140894808",gene="ATAD3B",
                     tissueSiteDetail="Adipose - Visceral (Omentum)",
                     datasetId="gtex_v7")

# Download expression using variant ID and gencode ID.
xQTLdownload_eqtlExp(variantName="chr1_14677_G_A_b38",gene="ENSG00000228463.9",
                    tissueSiteDetail="Stomach")
}
