% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_eqtlSig}
\alias{xQTLdownload_eqtlSig}
\title{xQTLdownload_eqtlSig}
\usage{
xQTLdownload_eqtlSig(
  variantName = "",
  genes = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{genes}{A gene symbol or a gencode id (versioned).}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{A character string. tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A data.table
}
\description{
download significant eQTL associations of a tissue or across all tissues.
}
\examples{
\donttest{
 # Download eQTL info for a variant:
 xQTLdownload_eqtlSig("rs201327123")
 xQTLdownload_eqtlSig("chr1_14677_G_A_b38")
 xQTLdownload_eqtlSig("11_66328719_T_C_b37", datasetId="gtex_v7")
 xQTLdownload_eqtlSig("11_66328719_T_C_b37", datasetId="gtex_v7",
                       tissueSiteDetail="Skin - Sun Exposed (Lower leg)")

 # Download eQTL association according to all tissues with genome location:
 varInfo <-  xQTLquery_varPos(chrom="chr1", pos=c(1102708),"gtex_v8")
 xQTLdownload_eqtlSig(variantName=varInfo$snpId)

 # Download eQTL info for gene:
 xQTLdownload_eqtlSig(genes="ATAD3B")
 xQTLdownload_eqtlSig(genes=c("TP53", "SLC35E2B"), tissueSiteDetail= "Brain - Cerebellum")
 xQTLdownload_eqtlSig(genes="ENSG00000141510.16", datasetId="gtex_v8")
 xQTLdownload_eqtlSig(genes="ENSG00000141510.11", datasetId="gtex_v7",
                      tissueSiteDetail="Thyroid" )

 # Download eQTL info for a variant-gene pair:
 xQTLdownload_eqtlSig(variantName="rs1641513", genes="TP53", datasetId="gtex_v8")
 xQTLdownload_eqtlSig(variantName="rs1641513", genes="TP53", datasetId="gtex_v7")
 xQTLdownload_eqtlSig(variantName="chr1_1667948_A_G_b38",
                   genes="SLC35E2B", tissueSiteDetail="Kidney - Cortex")
}
}
