% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_geneMedExp}
\alias{xQTLdownload_geneMedExp}
\title{xQTLdownload_geneMedExp}
\usage{
xQTLdownload_geneMedExp(
  genes = "",
  geneType = "auto",
  datasetId = "gtex_v8",
  tissueSiteDetail = "",
  recordPerChunk = 150
)
}
\arguments{
\item{genes}{A character vector of gene symbol/gencode ID.}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{"gtex_v7" or "gtex_v8"(default)}

\item{tissueSiteDetail}{A character string. Tissue detail can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{recordPerChunk}{1-2000. Default: 150.}
}
\value{
A data.frame
}
\description{
download genes' median expression in a tissue or across all tissues.
}
\examples{
\donttest{
 geneMedExp <- xQTLdownload_geneMedExp(genes="LYNX1")
 geneMedExp <- xQTLdownload_geneMedExp(genes=c("TP53", "IRF5"))
}
}
