% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_geneMedExp}
\alias{xQTLdownload_geneMedExp}
\title{Download median expression of all samples for specified genes across tissues.}
\usage{
xQTLdownload_geneMedExp(
  genes = "",
  geneType = "auto",
  datasetId = "gtex_v8",
  tissueSiteDetail = "",
  recordPerChunk = 150
)
}
\arguments{
\item{genes}{(character string or a character vector) gene symbols or gencode ids (versioned or unversioned are both supported).}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{(character) options: "gtex_v8" (default), "gtex_v7".}

\item{tissueSiteDetail}{(character) details of tissues in GTEx can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{recordPerChunk}{(integer) number of records fetched per request (default: 150).}
}
\value{
A data.table object.
}
\description{

}
\examples{
geneMedExp <- xQTLdownload_geneMedExp(genes="LYNX1")
geneMedExp <- xQTLdownload_geneMedExp(genes=c("TP53", "IRF5"))
}
