% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_ld}
\alias{xQTLdownload_ld}
\title{xQTLdownload_ld}
\usage{
xQTLdownload_ld(
  gene = "",
  geneType = "geneSymbol",
  datasetId = "gtex_v8",
  recordPerChunk = 300
)
}
\arguments{
\item{gene}{A gene symbol, gencode id (versioned), or a charater string of gene type.}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{A character string. Options: "gtex_v8" (default), "gtex_v7".}

\item{recordPerChunk}{A integer value (1-500). number of records fetched per request (default: 100).}
}
\value{
A data.frame
}
\description{
download linkage disequilibrium data of the variants associated with this gene.
}
\examples{
\donttest{
 xQTLdownload_ld("TP53" )
 xQTLdownload_ld("TP53", datasetId="gtex_v7")
 xQTLdownload_ld(gene="ENSG00000008128.22")
}
}
