% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_mQTL}
\alias{xQTLdownload_mQTL}
\title{download meta data of DNA methylation QTL (mQTL)}
\usage{
xQTLdownload_mQTL(
  cpg_id = "cg00000221",
  tissue_name = "WholeBlood",
  mQTL_meta = NULL
)
}
\arguments{
\item{cpg_id}{phenotype_id like: cg00000236, can be obtained using \code{xQTLdownload_mqtlmeta}}

\item{tissue_name}{(String)  One of the tissues: BreastMammaryTissue, ColonTransverse, KidneyCortex, Lung, MuscleSkeletal, Ovary, Prostate, Testis and WholeBlood}
}
\value{
A data.table object
}
\description{
download meta data of DNA methylation QTL (mQTL)
}
\examples{
\donttest{
mQTL_dt <- xQTLdownload_mQTL(cpg_id="cg00000221",tissue_name="Prostate")
}
}
