% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_mqtlmeta}
\alias{xQTLdownload_mqtlmeta}
\title{Download meta data of DNA methylation QTL (mQTL)}
\usage{
xQTLdownload_mqtlmeta(tissue_name = "BreastMammaryTissue")
}
\arguments{
\item{tissue_name}{(String)  One of the tissues: BreastMammaryTissue, ColonTransverse, KidneyCortex, Lung, MuscleSkeletal, Ovary, Prostate, Testis and WholeBlood}
}
\value{
A data.table object
}
\description{
Download meta data of DNA methylation QTL (mQTL)
}
\examples{
mQTL_meta<- xQTLdownload_mqtlmeta("Testis")
}
