% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{xQTLdownload_sgene}
\alias{xQTLdownload_sgene}
\title{xQTLdownload_sgene}
\usage{
xQTLdownload_sgene(
  gene = "",
  geneType = "auto",
  datasetId = "gtex_v8",
  tissueSiteDetail = "",
  recordPerChunk = 2000
)
}
\arguments{
\item{gene}{A charater string of gene symbol, gencode id (versioned). Can be null.}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{A character string. only support "gtex_v8". Default: "gtex_v8".}

\item{tissueSiteDetail}{A character string. Tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{recordPerChunk}{A integer value (1-2000). number of records fetched per request (default: 2000).}
}
\value{
a data.table
}
\description{
download sGenes (sQTL Genes).
sGenes are genes that have at least one significant sQTL acting upon them. Results may be filtered by tissue.
}
\examples{
\donttest{
 # don't run:
 #sGeneInfoAlltissue <- xQTLdownload_sgene()

 sGeneInfo <- xQTLdownload_sgene(tissueSiteDetail="Lung", recordPerChunk=2000)
 sGeneInfo <- xQTLdownload_sgene("ENSG00000141510.16",  tissueSiteDetail="Lung")
 sGeneInfo <- xQTLdownload_sgene("DDX11", tissueSiteDetail="Artery - Tibial" )
}
}
