% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_sqtlAllAsso}
\alias{xQTLdownload_sqtlAllAsso}
\title{download summary statistics of sQTL of a specified gene in GTEx v8.}
\usage{
xQTLdownload_sqtlAllAsso(
  genes = "",
  geneType = "gencodeId",
  tissue = "",
  clu_names = "",
  clu_geneid_DF = NULL
)
}
\arguments{
\item{genes}{(character) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "gencodeId".}

\item{tissue}{(character) details of tissues in GTEx can be listed using \code{tissueSiteDetailGTExv8}}

\item{clu_names}{(character) If provided, only the sQTL of clu_names will be downloaded}

\item{clu_geneid_DF}{(data.frame) If provided, clu-gencode mapping relationship will be loaded from this data.frame.}
}
\value{
A data.table object of sQTL dataset.
}
\description{
download summary statistics of sQTL of a specified gene in GTEx v8.
}
\examples{
\donttest{
sQTL_DT <- xQTLdownload_sqtlAllAsso(genes=c("MMP7","TP53"), tissue="Lung")
}
}
