% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{xQTLdownload_sqtlExp}
\alias{xQTLdownload_sqtlExp}
\title{xQTLdownload_sqtlExp}
\usage{
xQTLdownload_sqtlExp(
  variantName = "",
  phenotypeId = "",
  variantType = "snpId",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{phenotypeId}{Format like: "chr1:497299:498399:clu_54863:ENSG00000239906.1"}

\item{variantType}{A character string. "snpId" or "variantId". Default: "snpId".}

\item{tissueSiteDetail}{A character string.}

\item{datasetId}{A character string. Only GTEx v8 is supported, default: "gtex_v8".}
}
\value{
A data.table
}
\description{
download normalized expression of intron for a sQTL pair.
}
\examples{
\donttest{
 # Download exp in different tissues:
 xQTLdownload_sqtlExp(variantName="rs1450891501",
                      phenotypeId="chr1:497299:498399:clu_54863:ENSG00000239906.1",
                      tissueSiteDetail="Lung")

 # Dowload exp using variant ID and gencode ID.
 xQTLdownload_sqtlExp(variantName="chr1_1259424_T_C_b38",
                      variantType="variantId",
                      phenotypeId=" chr1:1487914:1489204:clu_52051:ENSG00000160072.19",
                      tissueSiteDetail="Adipose - Subcutaneous")
}
}
