% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExdownload.R
\name{xQTLdownload_sqtlSig}
\alias{xQTLdownload_sqtlSig}
\title{Fetch all eQTL associations from EBI eQTL category.}
\usage{
xQTLdownload_sqtlSig(
  variantName = "",
  gene = "",
  variantType = "snpId",
  geneType = "geneSymbol",
  tissueSiteDetail = ""
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{gene}{A gene symbol or a gencode id (versioned).}

\item{variantType}{A character string. "snpId" or "variantId". Default: "snpId".}

\item{geneType}{A character string. "geneSymbol"(default) or "gencodeId".}

\item{tissueSiteDetail}{A character string.
Tissue must be chosen from the following tissue names:
\tabular{rrrrr}{
\strong{tissue name} \tab \strong{GTEx V8} \cr
Adipose - Subcutaneous \tab √ \cr
Adipose - Visceral (Omentum) \tab √ \cr
Adrenal Gland \tab √ \cr
Artery - Aorta \tab √ \cr
Artery - Coronary \tab √ \cr
Artery - Tibial \tab √ \cr
Bladder \tab √ \cr
Brain - Amygdala \tab √ \cr
Brain - Anterior cingulate cortex (BA24) \tab √ \cr
Brain - Caudate (basal ganglia) \tab √ \cr
Brain - Cerebellar Hemisphere \tab √ \cr
Brain - Cerebellum \tab √ \cr
Brain - Cortex \tab √ \cr
Brain - Frontal Cortex (BA9) \tab √ \cr
Brain - Hippocampus \tab √ \cr
Brain - Hypothalamus \tab √ \cr
Brain - Nucleus accumbens (basal ganglia) \tab √ \cr
Brain - Putamen (basal ganglia) \tab √ \cr
Brain - Spinal cord (cervical c-1) \tab √ \cr
Brain - Substantia nigra \tab √ \cr
Breast - Mammary Tissue \tab √ \cr
Cells - Cultured fibroblasts \tab √ \cr
Cells - EBV-transformed lymphocytes \tab √ \cr
Cells - Transformed fibroblasts \tab x \cr
Cervix - Ectocervix \tab √ \cr
Cervix - Endocervix \tab √ \cr
Colon - Sigmoid \tab √ \cr
Colon - Transverse \tab √ \cr
Esophagus - Gastroesophageal Junction \tab √ \cr
Esophagus - Mucosa \tab √ \cr
Esophagus - Muscularis \tab √ \cr
Fallopian Tube \tab √ \cr
Heart - Atrial Appendage \tab √ \cr
Heart - Left Ventricle \tab √ \cr
Kidney - Cortex \tab √ \cr
Kidney - Medulla \tab √ \cr
Liver \tab √ \cr
Lung \tab √ \cr
Minor Salivary Gland \tab √ \cr
Muscle - Skeletal \tab √ \cr
Nerve - Tibial \tab √ \cr
Ovary \tab √ \cr
Pancreas \tab √ \cr
Pituitary \tab √ \cr
Prostate \tab √ \cr
Skin - Not Sun Exposed (Suprapubic) \tab √ \cr
Skin - Sun Exposed (Lower leg) \tab √ \cr
Small Intestine - Terminal Ileum \tab √ \cr
Spleen \tab √ \cr
Stomach \tab √ \cr
Testis \tab √ \cr
Thyroid \tab √ \cr
Uterus \tab √ \cr
Vagina \tab √ \cr
Whole Blood \tab √ \cr
}}

\item{refSeq}{A character string.}

\item{rsid}{dbSNP ID.}

\item{variantId}{A character string. like "chr17_5294580_CT_C_b38".}

\item{pThreshold}{Float. set a threshold for p-value to filter the result. Default: 1, aQTL associations that have a p-value<1 will be returned.}
}
\value{


A data.table
}
\description{
download significant sQTL associations of a tissue or across all tissues.
Only GTEx v8 is supported.
}
\examples{
\donttest{
  xQTLdownload_aQTLAllAsso( gene="TP53", tissueSiteDetail = "Adipose - Visceral (Omentum)" )
  xQTLdownload_aQTLAllAsso( refSeq = "NM_001291581.2", tissueSiteDetail = "Adipose - Visceral (Omentum)" )
  xQTLdownload_aQTLAllAsso( rsid = "rs3026133", tissueSiteDetail = "Adipose - Visceral (Omentum)", pThreshold=1e-1 )
  xQTLdownload_aQTLAllAsso( variantId = "chr17_5289661_A_G_b38", tissueSiteDetail = "Adipose - Visceral (Omentum)" )
}
\donttest{
 # Download sQTL info for a variant:
 xQTLdownload_sqtlSig(variantName="rs201327123", variantType="snpId")
 xQTLdownload_sqtlSig(variantName="chr1_14677_G_A_b38", variantType="variantId")
 xQTLdownload_sqtlSig(variantName="chr1_14677_G_A_b38", variantType="variantId",
                      tissueSiteDetail="Whole Blood")

 # Download sQTL association according to all tissues with genome location:
 varInfo <-  xQTLquery_varPos(chrom="chr1", pos=c(1102708),"gtex_v8")
 xQTLdownload_sqtlSig(variantName=varInfo$snpId, variantType="snpId")

 # Download sQTL info for a gene:
 xQTLdownload_sqtlSig(gene="ATAD3B")
 xQTLdownload_sqtlSig(gene="ENSG00000141510.16", geneType="gencodeId")
 xQTLdownload_sqtlSig(gene="ENSG00000141510.16", geneType="gencodeId",
                      tissueSiteDetail="Lung" )

 # Download eQTL info for a variant-gene pair:
 xQTLdownload_sqtlSig(variantName="rs546057177", gene="TP53")
 xQTLdownload_sqtlSig(variantName="chr17_7465085_A_G_b38", variantType="variantId",
                      gene="TP53", geneType="geneSymbol",
                      tissueSiteDetail="Lung")
}
}
