% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_sqtlSig}
\alias{xQTLdownload_sqtlSig}
\title{Download significant sQTL associations of a tissue or across all tissues}
\usage{
xQTLdownload_sqtlSig(
  variantName = "",
  genes = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = ""
)
}
\arguments{
\item{variantName}{(character) name of variant, dbsnp ID and variant id is supported, eg. "rs138420351" and "chr17_7796745_C_T_b38".}

\item{genes}{(character string or a character vector) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{variantType}{(character) options: "auto", "snpId" or "variantId". Default: "auto".}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{(character) details of tissues in GTEx can be listed using \code{tissueSiteDetailGTExv8} or \code{tissueSiteDetailGTExv7}}
}
\value{
A data.table object.
}
\description{
Only GTEx v8 is supported.
}
\examples{
# Download sQTL detail with rsid:
xQTLdownload_sqtlSig(variantName="rs201327123")
# xQTLdownload_sqtlSig(variantName="chr1_14677_G_A_b38")
# xQTLdownload_sqtlSig(variantName="chr1_14677_G_A_b38", tissueSiteDetail="Whole Blood")

# Download sQTL detail with gene symbol and gencode ID:
# xQTLdownload_sqtlSig(genes="ENSG00000141510.16", tissueSiteDetail="Lung" )
# xQTLdownload_sqtlSig(genes=c("ATAD3B", "MLH1"))
# xQTLdownload_sqtlSig(genes=c("ENSG00000141510.16","ENSG00000076242.14"))

# Download sQTL detail with the variant-genes pair:
# xQTLdownload_sqtlSig(variantName="rs1987191",genes="ATAD3B")
# xQTLdownload_sqtlSig(variantName="rs201327123", genes=c("WASH7P","RP11-206L10.2"))
#xQTLdownload_sqtlSig(variantName="chr17_7465085_A_G_b38", genes="TP53", tissueSiteDetail="Lung")
}
