% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{xQTLdownload_sqtlSig}
\alias{xQTLdownload_sqtlSig}
\title{xQTLdownload_sqtlSig}
\usage{
xQTLdownload_sqtlSig(
  variantName = "",
  gene = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = ""
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{gene}{A gene symbol or a gencode id (versioned).}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{A character string. tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}
}
\value{
A data.table
}
\description{
download significant sQTL associations of a tissue or across all tissues.
Only GTEx v8 is supported.
}
\examples{
\donttest{
 # Download sQTL detail for a variant:
 xQTLdownload_sqtlSig(variantName="rs201327123")
 xQTLdownload_sqtlSig(variantName="chr1_14677_G_A_b38")
 xQTLdownload_sqtlSig(variantName="chr1_14677_G_A_b38",
                      tissueSiteDetail="Whole Blood")

 # Download sQTL detail for a gene:
 xQTLdownload_sqtlSig(gene="ATAD3B")
 xQTLdownload_sqtlSig(gene="ENSG00000141510.16")
 xQTLdownload_sqtlSig(gene="ENSG00000141510.16",
                      tissueSiteDetail="Lung" )

 # Download eQTL detail for a variant-gene pair:
 xQTLdownload_sqtlSig(variantName="rs546057177", gene="TP53")
 xQTLdownload_sqtlSig(variantName="chr17_7465085_A_G_b38",
                      gene="TP53", tissueSiteDetail="Lung")
}
}
