% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{xQTLquery_gene}
\alias{xQTLquery_gene}
\title{xQTLquery_gene.}
\usage{
xQTLquery_gene(
  genes = "",
  geneType = "auto",
  gencodeVersion = "v26",
  recordPerChunk = 150
)
}
\arguments{
\item{genes}{A charater vector or a string of gene symbol, gencode id (versioned), or a charater string of gene type.
\itemize{
\item \strong{gene symbol (Default)}.

A character string or a character vector (case ignored). like: "tp53","naDK","SDF4".
\item \strong{gencode/ensemble id} (versioned or unversioned).

A character string or a character vector (case ignored). like: "ENSG00000210195.2","ENSG00000078808".
\item \strong{gene classification}.

when "geneType" is "geneCategory", supported "genes" can be listed using function \code{gencodeGenetype$V26} or \code{gencodeGenetype$V19}

}}

\item{geneType}{A character string. "auto"(default), "geneSymbol", "gencodeId" or "geneCategory".}

\item{gencodeVersion}{"v26" or "v19"}

\item{recordPerChunk}{A integer value. Defaulut: 150}
}
\value{
A data.table of queried gene information. including following columns:
\itemize{
\item \strong{genes.} Input genes
\item \strong{geneSymbol.} Gene symbol.
\item \strong{gencodeId.} Gencode/ensemble id (versioned).
\item \strong{entrezGeneId.} Entrez gene ID.
\item \strong{geneType.} Gene type.
\item \strong{chromosome.} Note: "chr" is added in gencode v26,
\item \strong{start.}
\item \strong{end.}
\item \strong{strand.}
\item \strong{tss.} Transcriptional start site.
\item \strong{gencodeVersion.} Gencode Version.
\item \strong{genomeBuild.} Genome version.
\item \strong{description.}
}
}
\description{
users can use this function to query basic information (including gene name, symbol, position and description, etc ) of genes.
}
\examples{
\donttest{
 # get all protein coding genes and description:
 protein_coding <- xQTLquery_gene(genes="protein coding", geneType="geneCategory", "v26" )
 # get all miRNA and description:
 miRNA <- xQTLquery_gene(genes="miRNA", geneType="geneCategory", "v19")

 # hg38 test:
 geneInfo <- xQTLquery_gene("TP53")
 geneInfo <- xQTLquery_gene(c("tp53","naDK","SDF4") )
 geneInfo <- xQTLquery_gene(c("ENSG00000210195.2","ENSG00000078808"),
                               geneType="gencodeId")

 # hg19 test:
 geneInfo <- xQTLquery_gene(c("TP53","naDK"),  gencodeVersion="v19")
 geneInfo <- xQTLquery_gene(c("ENSG00000141510.11","ENSG00000008130.11"),
                            geneType="gencodeId", gencodeVersion="v19")

 geneInfo <- xQTLquery_gene(unique(protein_coding$geneSymbol))
 }
}
