% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{xQTLquery_geneAll}
\alias{xQTLquery_geneAll}
\title{Fetch details of all genes supported in GTEx.}
\usage{
xQTLquery_geneAll(gencodeVersion = "v26", recordPerChunk = 2000)
}
\arguments{
\item{gencodeVersion}{(character) options: "v26"(default, matched with gtex_v8) or "v19"}

\item{recordPerChunk}{(integer) number of records fetched per request (default: 2000).}
}
\value{
A data.table object of all genes' information.
}
\description{
Fetch details of all genes supported in GTEx.
}
\examples{
# Fetch all genes. This will take several minutes:
# allGenes <- xQTLquery_geneAll()
}
