% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{xQTLquery_geneAll}
\alias{xQTLquery_geneAll}
\title{fetch all genes' info.}
\usage{
xQTLquery_geneAll(gencodeVersion = "v26", recordPerChunk = 2000)
}
\arguments{
\item{gencodeVersion}{A character string. "v26" or "v19".}

\item{recordPerChunk}{A integer value. From 1 to 2000.}
}
\value{
A data.table with all genes' information.
}
\description{
fetch all genes' info.
}
\examples{
\donttest{
  allGenes <- xQTLquery_geneAll()
  }
}
