% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{xQTLquery_tissue}
\alias{xQTLquery_tissue}
\title{Query details for a specified tissue}
\usage{
xQTLquery_tissue(tissueName = "")
}
\arguments{
\item{tissueName}{Tissue name, tissue ID or tissue site name. Default return all tissues' information. Can be choonse from \code{tissueSiteDetailGTExv8} or \code{tissueSiteDetailGTExv7}}
}
\value{
A data.table object.
}
\description{
Information includes tissue IDs, number of RNA-Seq samples, number of RNA-Seq samples with genotype, number of expressed genes, number of eGenes. Also includes tissueSiteDetail ID, name, abbreviation, uberon ID, and standard tissue colors. TissueSiteDetails are grouped by TissueSites. By default, this service reports from the latest GTEx release.
}
\examples{
tissueAll <- xQTLquery_tissue() # fetch all tissues
Brain <- xQTLquery_tissue("Brain")
}
