% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{xQTLquery_varId}
\alias{xQTLquery_varId}
\title{Query variant in GTEx with variant ID or dbSNP ID}
\usage{
xQTLquery_varId(variantName = "", variantType = "auto", datasetId = "gtex_v8")
}
\arguments{
\item{variantName}{(character) name of variant, dbsnp ID and variant id is supported, eg. "rs138420351" and "chr17_7796745_C_T_b38".}

\item{variantType}{(character) options: "auto", "snpId" or "variantId". Default: "auto".}

\item{datasetId}{(character) options: "gtex_v8" (default), "gtex_v7".}
}
\value{
A data.table.
}
\description{

}
\examples{
xQTLquery_varId("rs12596338")
xQTLquery_varId("rs12596338", datasetId="gtex_v7")
xQTLquery_varId("chr11_66561248_T_C_b38")
xQTLquery_varId("11_66328719_T_C_b37", variantType="variantId", datasetId="gtex_v7")
}
