% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{xQTLquery_varId}
\alias{xQTLquery_varId}
\title{Query variant information by snpId or variant ID.}
\usage{
xQTLquery_varId(variantName = "", variantType = "snpId", datasetId = "gtex_v8")
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{variantType}{A character string. "snpId" or "variantId". Default: "snpId".}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A data.table.
}
\description{
Query variant information by snpId or variant ID.
}
\examples{
 \donttest{
  xQTLquery_varId("rs12596338")
  xQTLquery_varId("rs12596338", datasetId="gtex_v7")
  xQTLquery_varId("chr11_66561248_T_C_b38", variantType="variantId", datasetId="gtex_v8")
  xQTLquery_varId("11_66328719_T_C_b37", variantType="variantId", datasetId="gtex_v7")
 }
}
