% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{xQTLquery_varPos}
\alias{xQTLquery_varPos}
\title{Query variants using genome position.}
\usage{
xQTLquery_varPos(chrom = "", pos = numeric(0), recordPerChunk = 200)
}
\arguments{
\item{chrom}{(character) name of chromesome, including chr1-chr22, chrX, chrY.}

\item{pos}{An integer array.}

\item{recordPerChunk}{(integer) number of records fetched per request (default: 200).}
}
\value{
A data.table object.
}
\description{
Query variants using genome position.
}
\examples{
xQTLquery_varPos(chrom="chr1", pos=c(1102708,1105739))
}
