% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExquery.R
\name{xQTLquery_varPos}
\alias{xQTLquery_varPos}
\title{xQTLquery_varPos}
\usage{
xQTLquery_varPos(
  chrom = "",
  pos = numeric(0),
  datasetId = "gtex_v8",
  recordPerChunk = 200
)
}
\arguments{
\item{chrom}{A character string.}

\item{pos}{An integer array.}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}

\item{recordPerChunk}{A integer value (1-200). number of records fetched per request (default: 200).}
}
\value{
A data.table.
}
\description{
query varints with genome postion.
}
\examples{
\donttest{
 xQTLquery_varPos(chrom="chr1", pos=c(1102708,1105739),"gtex_v8")
 xQTLquery_varPos(chrom="1", pos=c(1038088,1041119),"gtex_v7")
 xQTLquery_varPos("1", c(1246438, 1211944, 1148100),"gtex_v7")
}
}
