% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_anno}
\alias{xQTLvisual_anno}
\title{Visualizing annotated variants}
\usage{
xQTLvisual_anno(snpHits, pValueBy = 5, plotType = "bar")
}
\arguments{
\item{snpHits}{A data.table object from result of xQTLanno_genomic}

\item{pValueBy}{Cut step of pvlaue. Defaults: 5}

\item{plotType}{"bar" (Default), or "pie"}
}
\value{
A ggplot object
}
\description{
Visualizing annotated variants
}
\examples{
\donttest{
url1 <- "https://github.com/dingruofan/exampleData/raw/master/gwas/gwasSub.txt.gz"
snpInfo <- fread(url1, sep="\t")
snpHits <- xQTLanno_genomic(snpInfo)
xQTLvisual_anno(snpHits, plotType="bar")
}
}
