% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_anno}
\alias{xQTLvisual_anno}
\title{Visualizing annotated variants}
\usage{
xQTLvisual_anno(snpHits, pValueBy = 5, plotType = "point")
}
\arguments{
\item{snpHits}{A data.table object from result of xQTLanalyze_anno}

\item{pValueBy}{Cut step of pvlaue. Defaults: 5}

\item{plotType}{"point", "bar", or "pie"}
}
\value{
A ggplot object
}
\description{
Visualizing annotated variants
}
\examples{
\donttest{
url1 <- "http://github.com/dingruofan/exampleData/raw/master/gwas/gwasSub.txt.gz"
snpInfo <- fread(url1, sep="\t")
snpHits <- xQTLanalyze_anno(snpInfo)
xQTLvisual_anno(snpHits, plotType="point")
}
}
