% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_anno}
\alias{xQTLvisual_anno}
\title{Visualize enrichment of variants derived from \code{xQTLanno_genomic}}
\usage{
xQTLvisual_anno(snpHits, pValueBy = 5, annoType = "enrichment")
}
\arguments{
\item{snpHits}{A data.table object from result of xQTLanno_genomic}

\item{pValueBy}{Cut step of pvlaue. Defaults: 5}

\item{annoType}{"enrichment" or "overlapping"}
}
\value{
A ggplot object
}
\description{
Visualize enrichment of variants derived from \code{xQTLanno_genomic}
}
\examples{
\donttest{
url1 <- "http://bioinfo.szbl.ac.cn/xQTL_biolinks/xqtl_data/gwas/gwasSub.txt.gz"
snpInfo <- fread(url1, sep="\t")
snpHits <- xQTLanno_genomic(snpInfo)
p <- xQTLvisual_anno(snpHits)
}
}
