% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_eqtl}
\alias{xQTLvisual_eqtl}
\title{xQTLvisual_eqtl}
\usage{
xQTLvisual_eqtl(gene, geneType = "auto", datasetId = "gtex_v8")
}
\arguments{
\item{gene}{A gene symbol or a gencode id (versioned).}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A ggplot object
}
\description{
plot significance of all eQTL associations for a gene across tissues.
}
\examples{
\donttest{
  xQTLvisual_eqtl("KIF15")
  xQTLvisual_eqtl("MLH1")
}
}
