% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_eqtl}
\alias{xQTLvisual_eqtl}
\title{Box plot with jittered points for showing number and significance of eQTL associations}
\usage{
xQTLvisual_eqtl(gene, geneType = "auto")
}
\arguments{
\item{gene}{(character) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}
}
\value{
A ggplot object.
}
\description{
Box plot with jittered points for showing number and significance of eQTL associations
}
\examples{
\donttest{
xQTLvisual_eqtl("KIF15")
}
}
