% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_eqtlExp}
\alias{xQTLvisual_eqtlExp}
\title{xQTLvisual_eqtlExp}
\usage{
xQTLvisual_eqtlExp(
  variantName = "",
  gene = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{gene}{A gene symbol or a gencode id (versioned).}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{A character string. Tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A list containing expression profile and a ggplot object.
}
\description{
plot normalized expression among genotypes for eQTL.
}
\examples{
\donttest{
 # EQTL associatons of TP53:
 expEqtl <- xQTLvisual_eqtlExp(variantName="rs78378222", gene ="TP53",
                               tissueSiteDetail="Esophagus - Mucosa")
 expEqtl <- xQTLvisual_eqtlExp(variantName="rs78378222", gene ="TP53",
                               tissueSiteDetail="Lung")
 expEqtl <- xQTLvisual_eqtlExp(variantName="rs3778754", gene ="IRF5",
                               tissueSiteDetail="Whole Blood")
}
}
