% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_geneCorr}
\alias{xQTLvisual_geneCorr}
\title{Scatter plot for showing the correlation of two genes’ expression.}
\usage{
xQTLvisual_geneCorr(
  gene2 = "",
  geneType = "auto",
  tissueSiteDetail = "",
  groupBy = "sex",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{gene2}{Gene symbol or gencode ID of two genes. Default: gene symbol.}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissueSiteDetail}{(character) details of tissues in GTEx can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{groupBy}{Default:sex, can be choosen from pathologyNotesCategories, like: pathologyNotesCategories.mastopathy, pathologyNotesCategories.mastopathy.metaplasia.}

\item{datasetId}{(character) options: "gtex_v8" (default), "gtex_v7".}
}
\value{
A ggplot object.
}
\description{

}
\examples{
gene2 = c("AMOT", "HOXA4")
xQTLvisual_geneCorr(gene2,tissueSiteDetail="Liver")
xQTLvisual_geneCorr(gene2,groupBy="pathologyNotesCategories.congestion",tissueSiteDetail="Lung")
}
