% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_geneExpTissues}
\alias{xQTLvisual_geneExpTissues}
\title{xQTLvisual_geneExpTissues}
\usage{
xQTLvisual_geneExpTissues(
  gene = "",
  geneType = "geneSymbol",
  datasetId = "gtex_v8",
  toTissueSite = TRUE
)
}
\arguments{
\item{geneType}{"geneSymbol" or "gencodeId".}

\item{datasetId}{"gtex_v8" or "gtex_v7".}

\item{toTissueSite}{TRUE or FALSE, display all subtissues or tissue Site. Default: TURE.}

\item{genes}{A characer vector.}
}
\value{
A data.table and a plot.
}
\description{
plot distribution of the gene expression among multiple tissues.
}
\examples{
\donttest{
  geneExpTissues <- xQTLvisual_geneExpTissues("TP53",toTissueSite=TRUE)
}
}
