% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_geneExpTissues}
\alias{xQTLvisual_geneExpTissues}
\title{xQTLvisual_geneExpTissues}
\usage{
xQTLvisual_geneExpTissues(
  gene = "",
  geneType = "auto",
  datasetId = "gtex_v8",
  toTissueSite = TRUE
)
}
\arguments{
\item{gene}{A characer vector. Gene symbol or gencode Id.}

\item{geneType}{A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{datasetId}{"gtex_v8" or "gtex_v7". Default:"gtex_v8".}

\item{toTissueSite}{TRUE or FALSE, display all subtissues or tissue Site. Default: TURE.}
}
\value{
A list containing expression profile and a ggplot object.
}
\description{
plot distribution of the gene expression among multiple tissues.
}
\examples{
\donttest{
  geneExpTissues <- xQTLvisual_geneExpTissues("TP53",toTissueSite=TRUE)
}
}
