% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_geneExpTissues}
\alias{xQTLvisual_geneExpTissues}
\title{Violin plot of distribution of the gene expression profiles among multiple tissues.}
\usage{
xQTLvisual_geneExpTissues(
  gene = "",
  geneType = "auto",
  tissues = "All",
  log10y = FALSE,
  toTissueSite = FALSE,
  axis_text_size = 1.3,
  axis_title_size = 1.3,
  title_size = 1.4,
  xlab_text = "",
  ylab_text = "",
  title_text = ""
)
}
\arguments{
\item{gene}{(character) gene symbol or gencode id (versioned or unversioned are both supported).}

\item{geneType}{(character) options: "auto","geneSymbol" or "gencodeId". Default: "auto".}

\item{tissues}{A character string or a vector. "All" (default) means that all tissues is included.}

\item{log10y}{Display values of expression in log scale. Default: FALSE.}

\item{toTissueSite}{TRUE or FALSE, display all subtissues or tissue Site. Default: TURE.}

\item{axis_text_size}{(numberic) text size of the axis labels}

\item{axis_title_size}{(numberic) text size of the axis title}

\item{title_size}{(numberic) text size of the title of the plot}

\item{xlab_text}{(character) Lable for x-axis}

\item{ylab_text}{(character) Lable for x-axis}

\item{title_text}{(character) Title of the plot}
}
\value{
A list containing expression profile and a ggplot object.
}
\description{
Violin plot of distribution of the gene expression profiles among multiple tissues.
}
\examples{
\donttest{
# Display gene expression in specified tissues.
geneExpTissues <- xQTLvisual_geneExpTissues("TP53", tissues=c("Lung", "Brain","Ovary"))
}
}
