% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_genesExp}
\alias{xQTLvisual_genesExp}
\title{Density distribution of specified genes' expression profiles in a specified tissue.}
\usage{
xQTLvisual_genesExp(
  genes,
  geneType = "geneSymbol",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{genes}{Following gene types are supported:
\itemize{
\item \strong{Gene symbol}.

A character string or a character vector (case ignored). like: "tp53","naDK","SDF4".
\item \strong{Gencode/ensemble id} (versioned or unversioned).
}}

\item{geneType}{A character string. "geneSymbol"(default), "gencodeId" or "geneCategory".}

\item{tissueSiteDetail}{Tissue must be accessed by "tissueSiteDetailGTExv7" or "tissueSiteDetailGTExv7".}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A plot.
}
\description{
Density distribution of specified genes' expression profiles in a specified tissue.
}
\examples{
\donttest{
  genes <- c("FNDC8", "S100Z", "AQP6", "AMOT", "C3orf38", "FOXL1", "COX11", "FCN3", "DDX58", "CFI", "MS4A18", "NUDT13", "HOXA4", "VSX1")
  xQTLvisual_genesExp(genes, tissueSiteDetail="Lung")
  genes <- c("ENSG00000073598.5","ENSG00000171643.13","ENSG00000086159.12","ENSG00000126016.15","ENSG00000179021.9","ENSG00000176678.5","ENSG00000166260.10","ENSG00000142748.12","ENSG00000107201.9","ENSG00000205403.12","ENSG00000214782.7","ENSG00000166321.13","ENSG00000197576.13","ENSG00000100987.14")
  xQTLvisual_genesExp(genes, geneType="gencodeId", tissueSiteDetail="Liver")
}
}
