% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_locusCompare}
\alias{xQTLvisual_locusCompare}
\title{xQTLvisual_locusCompare}
\usage{
xQTLvisual_locusCompare(
  eqtlDF,
  gwasDF,
  highlightSnp = "",
  population = "EUR",
  token = "9246d2db7917",
  windowSize = 5e+05,
  genome = "grch38",
  snpLD = NA
)
}
\arguments{
\item{eqtlDF}{A data.frame or data.table with two columns: dbSNP id and p-value.}

\item{gwasDF}{A data.frame or data.table with two columns: dbSNP id and p-value.}

\item{highlightSnp}{Default is the snp that is farthest from the origin of the coordinates.}

\item{population}{Supported population is consistent with the LDlink, which can be listed using function LDlinkR::list_pop()}

\item{token}{LDlink provided user token, default = NULL, register for token at https://ldlink.nci.nih.gov/?tab=apiaccess}

\item{windowSize}{Window around the highlighted snp for querying linkage disequilibrium information. Default:500000}

\item{genome}{"grch38"(default) or "grch37".}
}
\value{
A plot
}
\description{
visualize the colocalization of association summary statistics.
}
\examples{
\donttest{
  eqtlURL <- "https://gitee.com/stronghoney/exampleData/raw/master/eqtl/eqtlAsso.txt"
  gwasURL <- "https://gitee.com/stronghoney/exampleData/raw/master/gwas/AD/gwasChr6Sub1.txt"
  eqtlDF <- data.table::fread(rawToChar(curl::curl_fetch_memory(eqtlURL)$content), sep="\t")
  gwasDF <- data.table::fread(rawToChar(curl::curl_fetch_memory(gwasURL)$content), sep="\t")
  eqtlDF <- eqtlDF[,.(snpId, pValue)]
  gwasDF <- gwasDF[,.(rsid, P)]
  xQTLvisual_locusCompare( eqtlDF, gwasDF )
}
}
