% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_PZPlot}
\alias{xQTLvisual_PZPlot}
\title{Compare P-values reported in the association result file to P-values calculated from Z statistics derived from the reported beta and standard error.}
\usage{
xQTLvisual_PZPlot(summaryDT)
}
\arguments{
\item{summaryDT}{A data.frame with three cols: pval,  beta, se.}
}
\value{
a list containing a data.frame of estimated pvalues and A ggplot2 object
}
\description{
Compare P-values reported in the association result file to P-values calculated from Z statistics derived from the reported beta and standard error.
}
\examples{
\donttest{
url1 <- "https://raw.githubusercontent.com/dingruofan/exampleData/master/eqtl/MMP7_qtlDF.txt"
qtl <- fread(url1, sep="\t")
xQTLanno_PZPlot(qtl[,.(pValue, beta, se)])
}
}
