% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{xQTLvisual_sqtlExp}
\alias{xQTLvisual_sqtlExp}
\title{xQTLvisual_sqtlExp}
\usage{
xQTLvisual_sqtlExp(
  variantName = "",
  phenotypeId = "",
  variantType = "auto",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)
}
\arguments{
\item{variantName}{A character string. like dbsnp ID or variant id in GTEx.}

\item{phenotypeId}{A character string. Format like: "chr1:497299:498399:clu_54863:ENSG00000239906.1"}

\item{variantType}{A character string. "auto", "snpId" or "variantId". Default: "auto".}

\item{tissueSiteDetail}{A character string. Tissue detail can be listed using "tissueSiteDetailGTExv8" or "tissueSiteDetailGTExv7"}

\item{datasetId}{A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".}
}
\value{
A list of variant detail and expression profile.
}
\description{
plot normalized expression among genotypes for sQTL.
}
\examples{
\donttest{
 # sQTL associatons of TP53:
 expSqtl <-xQTLvisual_sqtlExp(variantName="chr11_66561248_T_C_b38",
           phenotypeId ="chr11:66348070:66353455:clu_8500:ENSG00000255468.6",
           tissueSiteDetail="Skin - Sun Exposed (Lower leg)")
}
}
